Tarefa1

Description

Módulo para a realização da Tarefa 1 de LI1 em 2023/24.

Documentation

colisoesParede :: Mapa -> Personagem -> Bool

A função colisoesParede verifica se a personagem em questão colide com uma parede ou com uma borda do mapa.

Exemplos de utilização:

>>> colisoesParede Personagem
        { velocidade = (0,0)
        , tipo = Jogador
        , posicao = (0.26,5.5)
        , direcao = Oeste
        , tamanho = (0.5,1)
        , emEscada = False
        , ressalta = False
        , vida = 100
        , pontos = 100
        , aplicaDano = (False,0.0)
        } m1 = Mapa ((0.5, 5.5), Oeste) (0.5, 2.5) [[Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma]
                                                   ,[Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio]
                                                   ,[Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio]
                                                   ,[Plataforma, Plataforma, Vazio, Vazio, Vazio, Vazio, Plataforma, Plataforma, Plataforma, Plataforma]
                                                   ,[Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Escada, Vazio]
                                                   ,[Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Escada, Vazio]
                                                   ,[Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma]
                                                   ]
False

>>> colisoesParede Personagem
        { velocidade = (0,0)
        , tipo = Jogador
        , posicao = (0.25,5.5)
        , direcao = Oeste
        , tamanho = (0.5,1)
        , emEscada = False
        , ressalta = False
        , vida = 100
        , pontos = 100
        , aplicaDano = (False,0.0)
        } m1 = Mapa ((0.5, 5.5), Oeste) (0.5, 2.5) [[Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma]
                                                   ,[Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio]
                                                   ,[Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio]
                                                   ,[Plataforma, Plataforma, Vazio, Vazio, Vazio, Vazio, Plataforma, Plataforma, Plataforma, Plataforma]
                                                   ,[Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Escada, Vazio]
                                                   ,[Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Escada, Vazio]
                                                   ,[Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma]
                                                    ]
True

colisoesPersonagens :: Personagem -> Personagem -> Bool

A função colisoesPersonagens verifica se as personagens em questão colidem.

Exemplos de utilização:

>>> colisoesPersonagens Personagem
        { velocidade = (0,0)
        , tipo = Jogador
        , posicao = (1.01,5.5)
        , direcao = Oeste
        , tamanho = (0.5,1)
        , emEscada = False
        , ressalta = False
        , vida = 100
        , pontos = 100
        , aplicaDano = (False,0.0)
        } Personagem
        { velocidade = (0,0)
        , tipo = Jogador
        , posicao = (0.5,5.5)
        , direcao = Oeste
        , tamanho = (0.5,1)
        , emEscada = False
        , ressalta = False
        , vida = 100
        , pontos = 100
        , aplicaDano = (False,0.0)
        }
False

>>> colisoesPersonagens Personagem
        { velocidade = (0,0)
        , tipo = Jogador
        , posicao = (1,5.5)
        , direcao = Oeste
        , tamanho = (0.5,1)
        , emEscada = False
        , ressalta = False
        , vida = 100
        , pontos = 100
        , aplicaDano = (False,0.0)
        } Personagem
        { velocidade = (0,0)
        , tipo = Jogador
        , posicao = (0.5,5.5)
        , direcao = Oeste
        , tamanho = (0.5,1)
        , emEscada = False
        , ressalta = False
        , vida = 100
        , pontos = 100
        , aplicaDano = (False,0.0)
        }
True

checkProxParede :: Personagem -> Mapa -> Bool

A função checkProxBlocoParede verifica se a personagem em questão esté em colisão com um bloco de plataforma do mapa, excepto para as posições das margens (x e x tamanho do mapa e y < 1), que retorna sempre Vazio.

Exemplos de utilização:

>>> checkProxParede Personagem
        { velocidade = (0,0)
        , tipo = Jogador
        , posicao = (2.25,3)
        , direcao = Oeste
        , tamanho = (0.5,1.0)
        , emEscada = False
        , ressalta = False
        , vida = 100
        , pontos = 100
        , aplicaDano = (False,0.0)
        } Mapa ((0.5, 5.5), Oeste) (0.5, 2.5) [[Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma]
                                             ,[Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio]
                                             ,[Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio]
                                             ,[Plataforma, Plataforma, Vazio, Vazio, Vazio, Vazio, Plataforma, Plataforma, Plataforma, Plataforma]
                                             ,[Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Escada, Vazio]
                                             ,[Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Escada, Vazio]
                                             ,[Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma]
                                             ]
True

hitBoxBloco :: Posicao -> Hitbox

A função hitBoxBloco verifica a hitBox de um bloco numa certa posição.

Exemplos de utilização:

>>> hitBoxBloco (0,0)
((0.0,1.0),(1.0,0.0))

checkProxMargens :: Personagem -> Mapa -> Bool

A função checkProxMargens verifica se a personagem em questão está no limite esquerdo ou direito do mapa.

Exemplos de utilização:

>>> checkProxMargens Personagem
        { velocidade = (0,0)
        , tipo = Jogador
        , posicao = (0.26,5.5)
        , direcao = Oeste
        , tamanho = (0.5,1.0)
        , emEscada = False
        , ressalta = False
        , vida = 100
        , pontos = 100
        , aplicaDano = (False,0.0)
        } Mapa ((0.5, 5.5), Oeste) (0.5, 2.5) [[Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma]
                                             ,[Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio]
                                             ,[Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio]
                                             ,[Plataforma, Plataforma, Vazio, Vazio, Vazio, Vazio, Plataforma, Plataforma, Plataforma, Plataforma]
                                             ,[Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Escada, Vazio]
                                             ,[Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Escada, Vazio]
                                             ,[Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma]
                                             ]
False

>>> checkProxMargens Personagem
        { velocidade = (0,0)
        , tipo = Jogador
        , posicao = (0.25,5.5)
        , direcao = Oeste
        , tamanho = (0.5,1.0)
        , emEscada = False
        , ressalta = False
        , vida = 100
        , pontos = 100
        , aplicaDano = (False,0.0)
        } Mapa ((0.5, 5.5), Oeste) (0.5, 2.5) [[Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma]
                                             ,[Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio]
                                             ,[Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio]
                                             ,[Plataforma, Plataforma, Vazio, Vazio, Vazio, Vazio, Plataforma, Plataforma, Plataforma, Plataforma]
                                             ,[Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Escada, Vazio]
                                             ,[Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Escada, Vazio]
                                             ,[Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma]
                                             ]
True

hitBoxPersonagem :: Personagem -> Hitbox

A função hitBoxPersonagem calcula literalmente a hitbox da personagem em questão.

Exemplos de utilização:

>>> hitBoxPersonagem Personagem
        { velocidade = (0,0)
        , tipo = Jogador
        , posicao = (0.5,2.5)
        , direcao = Oeste
        , tamanho = (0.5,1.0)
        , emEscada = False
        , ressalta = False
        , vida = 100
        , pontos = 100
        , aplicaDano = (False,0.0)
        }
((0.25,2.0),(0.75,3.0))

intersectHitboxs :: Hitbox -> Hitbox -> Bool

A função intersectHitboxs verifica se existe algum tipo de interseção na hitbox de duas personagens distintas.

Exemplos de utilização:

>>> intersectHitboxs ((0.25,6.0),(0.75,5.0)) ((0.65,6.0),(1.15,5.0))
True

>>> intersectHitboxs ((0.25,6.0),(0.75,5.0)) ((1.25,6.0),(1.75,5.0))
False

tamanhoMapa :: Mapa -> (Double, Double)

A função tamanhoMapa calcula a altura e largura do mapa em questãoo através da matriz de blocos do mapa.

Exemplos de utilização:

>>> tamanhoMapa (Mapa ((0.5, 5.5), Oeste) (0.5, 2.5) [[Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma]
                                                    ,[Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio]
                                                    ,[Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio]
                                                    ,[Plataforma, Plataforma, Vazio, Vazio, Vazio, Vazio, Plataforma, Plataforma, Plataforma, Plataforma]
                                                    ,[Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Escada, Vazio]
                                                    ,[Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Escada, Vazio]
                                                    ,[Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma]
                                                    ]
(10.0,7.0)

aproxPerMatriz :: Personagem -> Posicao

A função aproxPerMatriz aproxima a posição da personagem à posição da matriz mais próxima.

Exemplos de utilizaçãoo:

>>> aproxPerMatriz = Personagem
        { velocidade = (0,0)
        , tipo = Jogador
        , posicao = (1.25,5.5)
        , direcao = Oeste
        , tamanho = (0.5,1)
        , emEscada = False
        , ressalta = False
        , vida = 100
        , pontos = 100
        , aplicaDano = (False,0.0)
        }
(1.0,5.0)

blocoNaMatriz :: Mapa -> Posicao -> Bloco

A função posicaoBloco indica qual tipo de bloco se situa na posicao em questão.

Exemplos de utilização:

>>> blocoNaMatriz (Mapa ((0.5, 5.5), Oeste) (0.5, 2.5) [[Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma]
                                                      ,[Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio]
                                                      ,[Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio]
                                                      ,[Plataforma, Plataforma, Vazio, Vazio, Vazio, Vazio, Plataforma, Plataforma, Plataforma, Plataforma]
                                                      ,[Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Escada, Vazio]
                                                      ,[Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Escada, Vazio]
                                                      ,[Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma]
                                                      ] (1,3)
Plataforma

>>> blocoNaMatriz (Mapa ((0.5, 5.5), Oeste) (0.5, 2.5) [[Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma]
                                                      ,[Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio]
                                                      ,[Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio]
                                                      ,[Plataforma, Plataforma, Vazio, Vazio, Vazio, Vazio, Plataforma, Plataforma, Plataforma, Plataforma]
                                                      ,[Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Escada, Vazio]
                                                      ,[Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Escada, Vazio]
                                                      ,[Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma]
                                                      ] (0,1)
Vazio