Tarefa2

Description

Módulo para a realização da Tarefa 2 de LI1 em 2023/24.

Documentation

valida :: Jogo -> Bool

A função valida verifica todos os parâmetros exigidos (no projeto) para um jogo.

Exemplos de utilização:

>>> valida Jogo
    { mapa = Mapa ((0.5, 5.5), Oeste) (0.5, 2.5) [[Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma]
                                                 ,[Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio]
                                                 ,[Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio]
                                                 ,[Plataforma, Plataforma, Vazio, Vazio, Vazio, Vazio, Plataforma, Plataforma, Plataforma, Plataforma]
                                                 ,[Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Escada, Vazio]
                                                 ,[Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Escada, Vazio]
                                                 ,[Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma]
                                                 ]
    , inimigos = [Personagem
        { velocidade = (0,0)
        , tipo = Fantasma
        , posicao = (0.5,2.5)
        , direcao = Oeste
        , tamanho = (0.5,1)
        , emEscada = False
        , ressalta = True
        , vida = 1
        , pontos = 100
        , aplicaDano = (False,0.0)
        }, Personagem
        { velocidade = (0,0)
        , tipo = Fantasma
        , posicao = (1.5,2.5)
        , direcao = Oeste
        , tamanho = (0.5,1)
        , emEscada = False
        , ressalta = True
        , vida = 1
        , pontos = 100
        , aplicaDano = (False,0.0)
        }, Personagem
        { velocidade = (0,0)
        , tipo = Fantasma
        , posicao = (2.5,2.5)
        , direcao = Oeste
        , tamanho = (0.5,1)
        , emEscada = False
        , ressalta = True
        , vida = 1
        , pontos = 100
        , aplicaDano = (False,0.0)
        }]
    , colecionaveis = [(Martelo,(2.5, 5.5)), (Moeda, (4.5, 5.5))]
    , jogador = Personagem
        { velocidade = (0,0)
        , tipo = Jogador
        , posicao = (1.75,5.5)
        , direcao = Oeste
        , tamanho = (2,1)
        , emEscada = False
        , ressalta = False
        , vida = 3
        , pontos = 100
        , aplicaDano = (False,0.0)
        }
    }
True

verifyChao :: Mapa -> Bool

A função verifyChao verifica se o mapa em questão tem um "chão".

Exemplos de utilização:

>>> verifyChao Mapa ((0.5, 5.5), Oeste) (0.5, 2.5) [[Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio]
                                                   ,[Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio]
                                                   ,[Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio]
                                                   ,[Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Escada, Vazio]
                                                   ,[Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Escada, Vazio]
                                                   ,[Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio]
                                                   ]
False

>>> verifyChao Mapa ((0.5, 5.5), Oeste) (0.5, 2.5) [[Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma]
                                                   ,[Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio]
                                                   ,[Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio]
                                                   ,[Plataforma, Plataforma, Vazio, Vazio, Vazio, Vazio, Plataforma, Plataforma, Plataforma, Plataforma]
                                                   ,[Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Escada, Vazio]
                                                   ,[Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Escada, Vazio]
                                                   ,[Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma]
                                                   ]
True

verifyResInimigos :: [Personagem] -> Bool

A função verifyResInimigos verifica se todos os inimigos têm a característica ressalta a True.

Exemplos de utilização:

>>> verifyResInimigos [Personagem
        { velocidade = (0,0)
        , tipo = Fantasma
        , posicao = (1.25,5.5)
        , direcao = Oeste
        , tamanho = (0.5,1)
        , emEscada = False
        , ressalta = True
        , vida = 100
        , pontos = 100
        , aplicaDano = (False,0.0)
        }, Personagem
        { velocidade = (0,0)
        , tipo = Fantasma
        , posicao = (1.5,5.5)
        , direcao = Oeste
        , tamanho = (0.5,1)
        , emEscada = False
        , ressalta = False
        , vida = 100
        , pontos = 100
        , aplicaDano = (False,0.0)
        }, Personagem
        { velocidade = (0,0)
        , tipo = Fantasma
        , posicao = (1.75,5.5)
        , direcao = Oeste
        , tamanho = (0.5,1)
        , emEscada = False
        , ressalta = True
        , vida = 100
        , pontos = 100
        , aplicaDano = (False,0.0)
        }]
False

verifyResPlayer :: Personagem -> Bool

A função verifyResPlayer verifica se o jogador tem a característica ressalta a False.

Exemplos de utilização:

>>> verifyResPlayer Personagem
        { velocidade = (0,0)
        , tipo = Jogador
        , posicao = (1.75,5.5)
        , direcao = Oeste
        , tamanho = (0.5,1)
        , emEscada = False
        , ressalta = False
        , vida = 100
        , pontos = 100
        , aplicaDano = (False,0.0)
        }
True

verifyPosicoesIni :: Mapa -> [Personagem] -> Bool

A função verifyPosicoesIni verifica se todos os inimigos têm a característica posicao diferente da inicial do jogador.

Exemplos de utilização:

>>> verifyPosicoesIni Mapa ((0.5, 5.5), Oeste) (0.5, 2.5) [[Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio]
                                                          ,[Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio]
                                                          ,[Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio]
                                                          ,[Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Escada, Vazio]
                                                          ,[Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Escada, Vazio]
                                                          ,[Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio]
                                                          ] 
        [Personagem
        { velocidade = (0,0)
        , tipo = Fantasma
        , posicao = (1.25,5.5)
        , direcao = Oeste
        , tamanho = (0.5,1)
        , emEscada = False
        , ressalta = True
        , vida = 100
        , pontos = 100
        , aplicaDano = (False,0.0)
        }, Personagem
        { velocidade = (0,0)
        , tipo = Fantasma
        , posicao = (1.5,5.5)
        , direcao = Oeste
        , tamanho = (0.5,1)
        , emEscada = False
        , ressalta = False
        , vida = 100
        , pontos = 100
        , aplicaDano = (False,0.0)
        }, Personagem
        { velocidade = (0,0)
        , tipo = Fantasma
        , posicao = (1.75,5.5)
        , direcao = Oeste
        , tamanho = (0.5,1)
        , emEscada = False
        , ressalta = True
        , vida = 100
        , pontos = 100
        , aplicaDano = (False,0.0)
        }]
True

verifyIniNumber :: [Personagem] -> Bool

A função verifyIniNumber verifica se o número de inimigos é superior a 2.

Exemplos de utilização:

>>> verifyIniNumber [Personagem
        { velocidade = (0,0)
        , tipo = Fantasma
        , posicao = (1.25,5.5)
        , direcao = Oeste
        , tamanho = (0.5,1)
        , emEscada = False
        , ressalta = True
        , vida = 100
        , pontos = 100
        , aplicaDano = (False,0.0)
        }, Personagem
        { velocidade = (0,0)
        , tipo = Fantasma
        , posicao = (1.5,5.5)
        , direcao = Oeste
        , tamanho = (0.5,1)
        , emEscada = False
        , ressalta = False
        , vida = 100
        , pontos = 100
        , aplicaDano = (False,0.0)
        }, Personagem
        { velocidade = (0,0)
        , tipo = Fantasma
        , posicao = (1.75,5.5)
        , direcao = Oeste
        , tamanho = (0.5,1)
        , emEscada = False
        , ressalta = True
        , vida = 100
        , pontos = 100
        , aplicaDano = (False,0.0)
        }]
True

verifyGhostLifes :: [Personagem] -> Bool

A função verifyGhostLifes verifica se todos os fantasmas têm exatamente uma vida.

Exemplos de utilização:

>>> verifyGhostLifes [Personagem
        { velocidade = (0,0)
        , tipo = Fantasma
        , posicao = (1.25,5.5)
        , direcao = Oeste
        , tamanho = (0.5,1)
        , emEscada = False
        , ressalta = True
        , vida = 1
        , pontos = 100
        , aplicaDano = (False,0.0)
        }, Personagem
        { velocidade = (0,0)
        , tipo = Fantasma
        , posicao = (1.5,5.5)
        , direcao = Oeste
        , tamanho = (0.5,1)
        , emEscada = False
        , ressalta = False
        , vida = 1
        , pontos = 100
        , aplicaDano = (False,0.0)
        }, Personagem
        { velocidade = (0,0)
        , tipo = Fantasma
        , posicao = (1.75,5.5)
        , direcao = Oeste
        , tamanho = (0.5,1)
        , emEscada = False
        , ressalta = True
        , vida = 1
        , pontos = 100
        , aplicaDano = (False,0.0)
        }]
True

verifyEscadas :: Mapa -> Posicao -> Bool

A função verifyEscadas verifica a composição de todas as escadas (respeitar as regras do projeto) do mapa em questão.

Exemplos de utilização:

>>> verifyEscadas m1 = Mapa ((0.5, 5.5), Oeste) (0.5, 2.5) [[Alcapao, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio]
                                                          ,[Escada, Vazio, Vazio, Vazio, Vazio, Vazio, Plataforma, Vazio, Vazio, Vazio]
                                                          ,[Escada, Vazio, Vazio, Vazio, Vazio, Vazio, Escada, Vazio, Vazio, Vazio]
                                                          ,[Plataforma, Vazio, Vazio, Vazio, Vazio, Vazio, Escada, Vazio, Vazio, Vazio]
                                                          ,[Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio]
                                                          ,[Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio]
                                                           ]
False

verifyEscadaComp :: Mapa -> Posicao -> Bool

A função verifyEscadaComp verifica a composição da escada (respeitar as regras do projeto) na posição mais próxima a '(x,y)' percorrendo a matriz da esquerda para a direita e de cima para baixo.

Exemplos de utilização:

>>> verifyEscadaComp m1 = Mapa ((0.5, 5.5), Oeste) (0.5, 2.5) [[Alcapao, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio]
                                                              ,[Escada, Vazio, Vazio, Vazio, Vazio, Vazio, Plataforma, Vazio, Vazio, Vazio]
                                                              ,[Escada, Vazio, Vazio, Vazio, Vazio, Vazio, Escada, Vazio, Vazio, Vazio]
                                                              ,[Plataforma, Vazio, Vazio, Vazio, Vazio, Vazio, Escada, Vazio, Vazio, Vazio]
                                                              ,[Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio]
                                                              ,[Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio]
                                                               ]
False

>>> verifyEscadaComp m1 = Mapa ((0.5, 5.5), Oeste) (0.5, 2.5) [[Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio]
                                                              ,[Escada, Vazio, Vazio, Vazio, Vazio, Vazio, Plataforma, Vazio, Vazio, Vazio]
                                                              ,[Escada, Vazio, Vazio, Vazio, Vazio, Vazio, Escada, Vazio, Vazio, Vazio]
                                                              ,[Plataforma, Vazio, Vazio, Vazio, Vazio, Vazio, Escada, Vazio, Vazio, Vazio]
                                                              ,[Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio]
                                                              ,[Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio]
                                                               ]
True

findEscada :: Mapa -> Posicao -> Posicao

A função findEscada indica a posição da primeira escada a ser encontrada a seguir à posição (x,y), percorrendo a matriz da esquerda para a direita e de cima para baixo.

Exemplos de utilização:

>>> findEscada [[Vazio, Vazio, Vazio, Plataforma, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio]
               ,[Vazio, Vazio, Vazio, Escada, Vazio, Vazio, Plataforma, Vazio, Vazio, Vazio]
               ,[Vazio, Vazio, Vazio, Escada, Vazio, Vazio, Escada, Vazio, Vazio, Vazio]
               ,[Vazio, Vazio, Vazio, Escada, Vazio, Vazio, Escada, Vazio, Vazio, Vazio]
               ,[Vazio, Vazio, Vazio, Escada, Vazio, Vazio, Plataforma, Vazio, Vazio, Vazio]
               ,[Vazio, Vazio, Vazio, Plataforma, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio]
                ]
(1.0,3.0)

blocoNaMatriz' :: Mapa -> Posicao -> Bloco

A função blocoNaMatriz' indica qual tipo de bloco se situa na posicao em questão.

Exemplos de utilização:

>>> blocoNaMatriz' (Mapa ((0.5, 5.5), Oeste) (0.5, 2.5) [[Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma]
                                                       ,[Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio]
                                                       ,[Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio]
                                                       ,[Plataforma, Plataforma, Vazio, Vazio, Vazio, Vazio, Plataforma, Plataforma, Plataforma, Plataforma]
                                                       ,[Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Escada, Vazio]
                                                       ,[Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Escada, Vazio]
                                                       ,[Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma]
                                                       ] (1,3)
Plataforma

>>> blocoNaMatriz' (Mapa ((0.5, 5.5), Oeste) (0.5, 2.5) [[Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma]
                                                       ,[Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio]
                                                       ,[Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio]
                                                       ,[Plataforma, Plataforma, Vazio, Vazio, Vazio, Vazio, Plataforma, Plataforma, Plataforma, Plataforma]
                                                       ,[Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Escada, Vazio]
                                                       ,[Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Escada, Vazio]
                                                       ,[Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma]
                                                       ] (0,1)
Vazio

verifyAlcapoesLength :: Mapa -> Personagem -> Bool

A função verifyAlcapoesLength verifica se todos os alcapoes incluídos no mapa não têm largura menor que a largura do jogador.

Exemplos de utilização:

>>> verifyAlcapoesLength (Mapa ((0.5, 5.5), Oeste) (0.5, 2.5) [[Alcapao, Alcapao, Alcapao, Alcapao, Alcapao, Alcapao, Alcapao, Alcapao, Alcapao, Alcapao]
                                                              ,[Alcapao, Alcapao, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio]
                                                              ,[Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio]
                                                              ,[Vazio, Alcapao, Alcapao, Alcapao, Alcapao, Vazio, Vazio, Vazio, Vazio, Vazio]
                                                              ,[Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio]
                                                              ,[Alcapao, Alcapao, Alcapao, Alcapao, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio]
                                                              ])
        (Personagem
        { velocidade = (0,0)
        , tipo = Jogador
        , posicao = (1.75,5.5)
        , direcao = Oeste
        , tamanho = (0.5,1)
        , emEscada = False
        , ressalta = False
        , vida = 100
        , pontos = 100
        , aplicaDano = (False,0.0)
        })
True

lengthAlcList :: Mapa -> Posicao -> [Double]

A função lengthAlcList calcula a lista das larguras de todos os Alcapoes do mapa.

Exemplos de utilização:

>>> lengthAlcList (Mapa ((0.5, 5.5), Oeste) (0.5, 2.5) [[Alcapao, Alcapao, Alcapao, Alcapao, Alcapao, Alcapao, Alcapao, Alcapao, Alcapao, Alcapao]
                                                       ,[Alcapao, Alcapao, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio]
                                                       ,[Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio]
                                                       ,[Vazio, Alcapao, Alcapao, Alcapao, Alcapao, Vazio, Vazio, Vazio, Vazio, Vazio]
                                                       ,[Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio]
                                                       ,[Alcapao, Alcapao, Alcapao, Alcapao, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio]
                                                       ]) (0,0)
[10.0,2.0,4.0,4.0]

lengthPlayer :: Personagem -> Double

A função lengthPlayer calcula a largura do jogador arredonda à unidade acima.

Exemplos de utilização:

>>> lengthPlayer Personagem
        { velocidade = (0,0)
        , tipo = Jogador
        , posicao = (1.75,5.5)
        , direcao = Oeste
        , tamanho = (0.5,1)
        , emEscada = False
        , ressalta = False
        , vida = 100
        , pontos = 100
        , aplicaDano = (False,0.0)
        }
1.0

verifyAlcapaoComp :: Mapa -> Posicao -> Double

A função verifyAlcapaoComp verifica a composição do alcapao (respeitar as regras do projeto) na posição mais próxima a '(x,y)' percorrendo a matriz da esquerda para a direita e de cima para baixo.

Exemplos de utilização:

>>> verifyAlcapaoComp Mapa ((0.5, 5.5), Oeste) (0.5, 2.5) [[Alcapao, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio]
                                                          ,[Escada, Vazio, Vazio, Vazio, Vazio, Vazio, Plataforma, Vazio, Vazio, Vazio]
                                                          ,[Escada, Vazio, Vazio, Vazio, Vazio, Vazio, Escada, Vazio, Vazio, Vazio]
                                                          ,[Plataforma, Vazio, Vazio, Vazio, Vazio, Vazio, Escada, Vazio, Vazio, Vazio]
                                                          ,[Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio]
                                                          ,[Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio]
                                                           ] (0,0)
1.0

>>> verifyAlcapaoComp Mapa ((0.5, 5.5), Oeste) (0.5, 2.5) [[Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio]
                                                          ,[Escada, Vazio, Vazio, Vazio, Vazio, Vazio, Plataforma, Vazio, Vazio, Vazio]
                                                          ,[Escada, Vazio, Alcapao, Alcapao, Vazio, Vazio, Escada, Vazio, Vazio, Vazio]
                                                          ,[Plataforma, Vazio, Vazio, Vazio, Vazio, Vazio, Escada, Vazio, Vazio, Vazio]
                                                          ,[Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio]
                                                          ,[Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio]
                                                           ] (0,0)
2.0

findAlcapao :: Mapa -> Posicao -> Posicao

A função findAlcapao indica a posição do primeiro alcapao a ser encontrado a seguir à posição (x,y), percorrendo a matriz da esquerda para a direita e de cima para baixo.

Exemplos de utilização:

>>> findAlcapao [[Vazio, Vazio, Vazio, Plataforma, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio]
               ,[Vazio, Vazio, Vazio, Escada, Vazio, Vazio, Plataforma, Vazio, Vazio, Vazio]
               ,[Vazio, Vazio, Alcapao, Escada, Vazio, Vazio, Escada, Vazio, Vazio, Vazio]
               ,[Vazio, Vazio, Vazio, Escada, Vazio, Vazio, Escada, Vazio, Vazio, Vazio]
               ,[Vazio, Vazio, Vazio, Escada, Vazio, Vazio, Plataforma, Vazio, Vazio, Vazio]
               ,[Vazio, Vazio, Vazio, Plataforma, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio]
                ]
(2.0,2.0)

verifyColPerInside :: Jogo -> Bool

A função verifyColPerInside verifica se todos os colecionáveis, inimigos e jogador se encontram num bloco do tipo Vazio.

Exemplos de utilização:

>>> verifyColPerInside Jogo
    { mapa = Mapa ((0.5, 5.5), Oeste) (0.5, 2.5) [[Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma]
                                                 ,[Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio]
                                                 ,[Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio]
                                                 ,[Plataforma, Plataforma, Vazio, Vazio, Vazio, Vazio, Plataforma, Plataforma, Plataforma, Plataforma]
                                                 ,[Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Escada, Vazio]
                                                 ,[Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Escada, Vazio]
                                                 ,[Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma]
                                                 ]
    , inimigos = [Personagem
        { velocidade = (0,0)
        , tipo = Fantasma
        , posicao = (0.5,2.5)
        , direcao = Oeste
        , tamanho = (0.5,1)
        , emEscada = False
        , ressalta = True
        , vida = 1
        , pontos = 100
        , aplicaDano = (False,0.0)
        }, Personagem
        { velocidade = (0,0)
        , tipo = Fantasma
        , posicao = (1.5,2.5)
        , direcao = Oeste
        , tamanho = (0.5,1)
        , emEscada = False
        , ressalta = True
        , vida = 1
        , pontos = 100
        , aplicaDano = (False,0.0)
        }, Personagem
        { velocidade = (0,0)
        , tipo = Fantasma
        , posicao = (2.5,2.5)
        , direcao = Oeste
        , tamanho = (0.5,1)
        , emEscada = False
        , ressalta = True
        , vida = 1
        , pontos = 100
        , aplicaDano = (False,0.0)
        }]
    , colecionaveis = [(Martelo,(2.5, 5.5)), (Moeda, (4.5, 5.5))]
    , jogador = Personagem
        { velocidade = (0,0)
        , tipo = Jogador
        , posicao = (1.75,5.5)
        , direcao = Oeste
        , tamanho = (2,1)
        , emEscada = False
        , ressalta = False
        , vida = 3
        , pontos = 100
        , aplicaDano = (False,0.0)
        }
    }
True

verifyColInside :: Mapa -> [(Colecionavel, Posicao)] -> Bool

A função verifyColInside verifica se todos os colecionáveis se encontram num bloco Vazio.

Exemplos de utilização:

>>> verifyColInside (Mapa ((0.5, 5.5), Oeste) (0.5, 2.5) [[Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma]
                                                         ,[Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio]
                                                         ,[Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio]
                                                         ,[Plataforma, Plataforma, Vazio, Vazio, Vazio, Vazio, Plataforma, Plataforma, Plataforma, Plataforma]
                                                         ,[Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Escada, Vazio]
                                                         ,[Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Escada, Vazio]
                                                         ,[Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma]
                                                         ]) [(Martelo,(2.5, 5.5)), (Moeda, (4.5, 5.5))]
True

blocoNaMatriz'' :: Mapa -> Posicao -> Bloco

A função blocoNaMatriz'' indica qual tipo de bloco se situa na posicao em questão, adptada para as funções inside.

Exemplos de utilização:

>>> blocoNaMatriz (Mapa ((0.5, 5.5), Oeste) (0.5, 2.5) [[Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma]
                                                      ,[Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio]
                                                      ,[Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio]
                                                      ,[Plataforma, Plataforma, Vazio, Vazio, Vazio, Vazio, Plataforma, Plataforma, Plataforma, Plataforma]
                                                      ,[Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Escada, Vazio]
                                                      ,[Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Escada, Vazio]
                                                      ,[Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma]
                                                      ] (1,3)
Plataforma

>>> blocoNaMatriz (Mapa ((0.5, 5.5), Oeste) (0.5, 2.5) [[Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma]
                                                      ,[Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio]
                                                      ,[Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio]
                                                      ,[Plataforma, Plataforma, Vazio, Vazio, Vazio, Vazio, Plataforma, Plataforma, Plataforma, Plataforma]
                                                      ,[Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Escada, Vazio]
                                                      ,[Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Escada, Vazio]
                                                      ,[Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma]
                                                      ] (0,1)
Vazio

verifyIsInside :: Mapa -> [Personagem] -> Bool

A função verifyIsInside verifica se todos os inimigos se encontram num bloco Vazio.

Exemplos de utilização:

>>> verifyIsInside (Mapa ((0.5, 5.5), Oeste) (0.5, 2.5) [[Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma]
                                                       ,[Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio]
                                                       ,[Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio]
                                                       ,[Plataforma, Plataforma, Vazio, Vazio, Vazio, Vazio, Plataforma, Plataforma, Plataforma, Plataforma]
                                                       ,[Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Escada, Vazio]
                                                       ,[Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Escada, Vazio]
                                                       ,[Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma]
                                                       ]) [Personagem
        { velocidade = (0,0)
        , tipo = Fantasma
        , posicao = (0.5,2.5)
        , direcao = Oeste
        , tamanho = (0.5,1)
        , emEscada = False
        , ressalta = True
        , vida = 1
        , pontos = 100
        , aplicaDano = (False,0.0)
        }, Personagem
        { velocidade = (0,0)
        , tipo = Fantasma
        , posicao = (1.5,2.5)
        , direcao = Oeste
        , tamanho = (0.5,1)
        , emEscada = False
        , ressalta = True
        , vida = 1
        , pontos = 100
        , aplicaDano = (False,0.0)
        }, Personagem
        { velocidade = (0,0)
        , tipo = Fantasma
        , posicao = (2.5,2.5)
        , direcao = Oeste
        , tamanho = (0.5,1)
        , emEscada = False
        , ressalta = True
        , vida = 1
        , pontos = 100
        , aplicaDano = (False,0.0)
        }]
True

verifyPlayerInside :: Mapa -> Personagem -> Bool

A função verifyPlayerInside verifica se o jogador se encontra num bloco Vazio.

Exemplos de utilização:

>>> verifyPlayerInside (Mapa ((0.5, 5.5), Oeste) (0.5, 2.5) [[Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma]
                                                            ,[Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio]
                                                            ,[Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio]
                                                            ,[Plataforma, Plataforma, Vazio, Vazio, Vazio, Vazio, Plataforma, Plataforma, Plataforma, Plataforma]
                                                            ,[Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Escada, Vazio]
                                                            ,[Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Escada, Vazio]
                                                            ,[Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma, Plataforma]
                                                            ]) 
        Personagem
        { velocidade = (0,0)
        , tipo = Jogador
        , posicao = (1.75,5.5)
        , direcao = Oeste
        , tamanho = (2,1)
        , emEscada = False
        , ressalta = False
        , vida = 3
        , pontos = 100
        , aplicaDano = (False,0.0)
        }
True