Tarefa3

Description

Módulo para a realização da Tarefa 3 de LI1 em 2023/24.

Documentation

movimenta :: Semente -> Tempo -> Jogo -> Jogo

Nesta tarefa, temos como objetivo definir a função movimenta, cujo mesma deve animar todos os personagens, i.e calcular as suas novas posiçõoes e respectivas consequências.

Para a definir, precisamos usar as seguintes funções auxiliares:

efeitoGravidade, que dado um jogo, aplica o vetor gravidade a todos os personagens do jogo;

animaPers, que dado um jogo, retorna um jogo, onde as posições de todos os personagens foram atualizadas;

Para esta função usamos a função auxiliar: animaIndividual, responsável por atualizar a posição de um personagem individual.

tiposColisoes, que dado um jogo, retorna um jogo, calculando as respetivas consequências de diferentes tipos de colisões.

Para definirmos esta função auxiliar, utilizamos outras funções, uma para cada tipo de colisão: colisoesPersPers, que trata das colisões entre dois personagens, atualizando os seus respetivos pontos de vida; colisoesPersMapa, que que atualiza as posições dos personagens e o mapa dependendo de certas colisões e colisoesJogCol, que atualiza a lista de colecionáveis de um Jogo, assim como os pontos do jogador.

Optamos por definir outras funções para facilitar o nosso trabalho no decorrer da tarefa:

criaHitboxDano, que retorna a hitbox de dano de um jogador quando está com o martelo ativado;

hitboxColecionavel, que retona a hitbox de um colecionável dada a respetiva posição;

hitboxTipoBloco, análoga à função hitboxColecionavel;

desapareceAlcapao, que retorna True se a hitbox de um jogador sobrepor a de um alcapão, ou seja se um jogador pisar um alcapão.

Funções:

Função Principal:

movimenta :: Semente -> Tempo -> Jogo -> Jogo

movimenta semente t jogo = jogoComColisoes

where random = geraAleatorios semente 10

jogoGravidade = efeitoGravidade jogo

jogoAnimado = animaPers random t jogoGravidade

jogoComColisoes = tiposColisoes jogoAnimado

Função auxiliar efeitoGravidade:

efeitoGravidade :: Jogo -> Jogo

efeitoGravidade jogo = jogo { inimigos = inimigosQueda, jogador = jogadorQueda }

where inimigosQueda = map (\inimigo -> inimigo { velocidade = sumVetor (velocidade inimigo) gravidade }) (inimigos jogo)

jogadorQueda = (jogador jogo) { velocidade = sumVetor (velocidade (jogador jogo)) gravidade }

Função auxiliar animaPers:

animaPers :: [Int] -> Tempo -> Jogo -> Jogo

animaPers random t jogo = jogo { inimigos = inimigosAnimados, jogador = jogadorAnimado }

where inimigosAnimados = map (\inimigo -> animaIndividual random t inimigo) (inimigos jogo)

jogadorAnimado = animaIndividual random t (jogador jogo)

animaIndividual :: [Int] -> Tempo -> Personagem -> Personagem

animaIndividual random t personagem = personagemAtualizado

where (vx, vy) = velocidade personagem

(x, y) = posicao personagem

novaPos = (x + vx * t, y + vy * t)

personagemAtualizado = case (tipo personagem) of

MacacoMalvado -> personagem

Fantasma -> personagem { posicao = novaPos }

Jogador -> personagem { posicao = novaPos }

Função auxiliar tipoColisoes:

tiposColisoes :: Jogo -> Jogo

tiposColisoes jogo = jogoComColisoesJogCol

where jogoComColisoesPersPers = colisoesPersPers jogo

jogoComColisoesPersMapa = colisoesPersMapa jogoComColisoesPersPers

jogoComColisoesJogCol = colisoesJogCol jogoComColisoesPersMapa

Função auxiliar colisoesPersPers:

colisoesPersPers :: Jogo -> Jogo

colisoesPersPers jogo = jogo { jogador = jogadorNovo , inimigos = inimigosAtualizados  }

where jogadorNovo = foldl (\jog inimigo -> colisaoJog jog inimigo) (jogador jogo) (inimigos jogo)

inimigosAtualizados = colisaoInimigos (jogador jogo) (inimigos jogo)

Função auxiliar colisoesPersMapa:

colisoesPersMapa :: Jogo -> Jogo

colisoesPersMapa jogo = jogo { jogador = novoJogador, inimigos = novosInimigos, mapa = novoMapa }

where novosInimigos = intersetamMapa (mapa jogo) (inimigos jogo)

novoJogador = intersetaMapa (mapa jogo) (jogador jogo)

novoMapa = pisaBlocosAlcp novoJogador (mapa jogo)

Função auxiliar colisoesJogCol:

colisoesJogCol :: Jogo -> Jogo

colisoesJogCol jogo = jogo { colecionaveis = colecionaveisAtualizados, jogador = jogadorNovo }

where (colecionaveisAtualizados, colecionaveisObtidos) = foldl (\(colAtualiza, recolhidos) (col, coord) ->

if colisaoHitHit (criaHitbox (jogador jogo)) (hitboxColecionavel coord)

then (colAtualiza, col : recolhidos)

else ((col, coord) : colAtualiza, recolhidos)) ([], []) (colecionaveis jogo)

jogadorNovo = foldl (\jog col -> apanhaCol jog col) (jogador jogo) colecionaveisObtidos

apanhaCol :: Personagem -> Colecionavel -> Personagem

apanhaCol jog col = case col of

Moeda -> jog { pontos = (pontos jog) + 1 }

Martelo -> jog { aplicaDano = (True, 10) }

criaHitboxDano :: Personagem -> Hitbox

Função que obtêm a hitBox de dano do jogador em função da sua posição e tamanho.

hitboxColecionavel :: Posicao -> Hitbox

Cria uma hitbox para um colecionável.

hitboxTipoBloco :: Posicao -> Hitbox

Cria uma hitbox para um bloco.

hitPosicao :: Hitbox -> Posicao

Transforma hitbox em posição.

desapareceAlcapao :: Hitbox -> Hitbox -> Bool

Retorna True se o jogador pisar um alcapao.

sumVetor :: Velocidade -> Velocidade -> Velocidade

Função que soma dois vetores.

efeitoGravidade :: Jogo -> Jogo

Função que aplica o vetor gravidade ao jogo.

animaPers :: [Int] -> Tempo -> Jogo -> Jogo

Função que calcula novas posições dos personagens.

animaIndividual :: [Int] -> Tempo -> Personagem -> Personagem

Função auxiliar de animaPers.

tiposColisoes :: Jogo -> Jogo

Função que trata das colisões.

colisoesPersPers :: Jogo -> Jogo

Função para tratar de colisões Inimigo-Jogador.

colisaoInimigos :: Personagem -> [Personagem] -> [Personagem]

Função auxiliar de colisoesPersPers.

colisaoJog :: Personagem -> Personagem -> Personagem

Função auxiliar de colisoesPersPers.

colisoesPersMapa :: Jogo -> Jogo

Função para tratar de colisões Personagem-Mapa.

intersetamMapa :: Mapa -> [Personagem] -> [Personagem]

Função auxiliar de colisoesPersMapa.

intersetaMapa :: Mapa -> Personagem -> Personagem

Função auxiliar de colisoesPersMapa.

passaLimites :: Hitbox -> Mapa -> Posicao

Função para resolver colisoes Limites-Personagem.

auxPassaLimites :: [[Bloco]] -> Hitbox -> Hitbox

Função auxiliar de passaLimites.

passaBlocos :: Hitbox -> Mapa -> Posicao

Função para resolver colisões Personagem-Blocos(Plataforma/Alcapão)

aux1PassaBlocos :: Hitbox -> [[Bloco]] -> Bool

Função auxiliar de passaBlocos.

aux2PassaBlocos :: Hitbox -> [[Bloco]] -> Posicao

Função auxiliar de passaBlocos.

aux3PassaBlocos :: Hitbox -> Posicao -> Posicao -> Posicao

Função auxiliar de passaBlocos.

pisaBlocosAlcp :: Personagem -> Mapa -> Mapa

Função que transforma Alcapao em Vazio caso o jogador o pise

eJog :: Personagem -> Bool

Função que testa se um personagem é do tipo Jogador.

hitboxMatriz' :: [[Bloco]] -> [Hitbox]

Função auxiliar de pisaBlocosAlcp.

hitboxLinha' :: (Int, [Bloco]) -> [Hitbox]

Função auxiliar de hitboxMatriz'.

hitboxBloco' :: Int -> Int -> Bloco -> [Hitbox]

Função auxiliar de hitboxMatriz'.

aux1PassaBlocosAlcp :: Hitbox -> [[Bloco]] -> Bool

Função auxiliar de pisaBlocosAlcp.

aux2PassaBlocosAlcp :: Personagem -> [[Bloco]] -> [[Bloco]]

Função auxiliar de pisaBlocosAlcp.

trocaLinha :: Hitbox -> (Int, [Bloco]) -> [Bloco]

Função auxiliar de pisaBlocosAlcp.

trocaBloco :: Int -> Hitbox -> (Int, Bloco) -> Bloco

Função auxiliar de pisaBlocosAlcp.

colisoesJogCol :: Jogo -> Jogo

Função para tratar de colisões Jogador-Colecionável.

apanhaCol :: Personagem -> Colecionavel -> Personagem

Função auxiliar de colisoesJogCol.