Tarefa5

Description

Este módulo contém as funcões que auxiliam a parte gráfica, tais como as funcões que reagem aos eventos, ao tempo, desenha, etc...

Documentation

sc :: Mapa -> Picture -> Picture

"sc" dá scale ás imagens do jogo

escalaJogo :: Mapa -> Float

"escalaJogo" define a escala em que vai ser redimensionado o jogo.

tamanhoBlocoW :: Mapa -> Float

"tamanhoBlocoW" define o tamanho que cada bloco vai ter na janela

desenhaMapa :: Mapa -> Double -> Mapa -> [Picture] -> [Picture]

"desenhaMapa" desenha o mapa do jogo

desenhaLinhaMapa :: [Bloco] -> Posicao -> Mapa -> [Picture] -> [Picture]

"desenhaLinhaMapa" desenha linha a linha o mapa do jogo

desenhaPlayer :: Picture -> [Picture]

"desenhaPlayer" desenha o jogador do jogo

desenhaPlayerAux :: Personagem -> Mapa -> [Picture] -> Picture

"desenhaPlayerAux" ajuda a funcao principal a desenhar o jogador

desenhaInimigos :: [Personagem] -> Mapa -> [Picture] -> [Picture]

"desenhaInimigos" desenha os inimigos existentes no jogo

desenhaInimigo :: Personagem -> Mapa -> [Picture] -> Picture

"desenhaInimigo" ajuda a funcao principal a desenhar os inimigos

desenhaColecionaveis :: [(Colecionavel, Posicao)] -> Mapa -> [Picture] -> [Picture]

"desenhaColecionaveis" desenha os colecionáveis existentes no jogo

desenhaColecionavel :: (Colecionavel, Posicao) -> Mapa -> [Picture] -> Picture

"desenhaColecionavel" ajuda a funcao principal a desenhar os colecionáveis

desenhaStar :: Mapa -> [Picture] -> [Picture]

"desenhaStar" desenha a estrela que indica a saída do jogo