Tarefa1

Description

Módulo para a realização da Tarefa 1 de LI1 em 2023/24.

Documentation

colisoesParede :: Mapa -> Personagem -> Bool

Nesta tarefa, temos como objetivo definir duas funçẽs:

colisoesParede que verifica se um personagem ultrapassa os limites do mapa ou colide com certos blocos do mapa;

Para definir esta função fizemos uso de duas funções auxiliares, sendo elas: colisaoHitboxparede, que teste se a hitbox de um personagem colide com as hitboxes de certos blocos do mapa, e testePassalimite, que testa se a hitbox de um personagem passa dos limites ou sai fora do mapa.

colisoesPersonagens que verifica se dois personagens estão em colisão.

Para definir esta função usamos uma função auxiliar: colisaoHitHit, que testa se duas hitboxes quaisqueres colidem.

Para obtermos as hitboxes necessárias para definir as funções principais usamos outras duas funções auxiliares:

criaHitbox, que dado um personagem cria a hitbox correspondente;

hitboxMatriz, que dado uma matriz de blocos, nomeadamente o mapa, dá as hitboxes dos blocos do tipo Plataforma e Alcapao.

Funções:

Funções Principais:

colisoesParede :: Mapa -> Personagem -> Bool

colisoesParede (Mapa _ _ l) p = colisaoHitboxparede hitboxp lp || testePassalimite l hitboxp

where hitboxp = criaHitbox p

lp = hitboxMatriz l

colisoesPersonagens :: Personagem -> Personagem -> Bool

colisoesPersonagens p1 p2 = colisaoHitHit hitboxp1 hitboxp2

where hitboxp1 = criaHitbox p1

hitboxp2 = criaHitbox p2

Função auxiliar criaHitbox:

criaHitbox :: Personagem -> Hitbox

criaHitbox personagem = ((x-l/2, y+c/2), (x+l/2, y-c/2))

where (x, y) = posicao personagem

(c, l) = tamanho personagem

Função auxiliar testePassalimite:

testePassalimite :: [[Bloco]] -> Hitbox -> Bool

testePassalimite l ((x1,y1),(x2,y2)) = x1 < 0 || x2 > nColunas || y1 > nLinhas || y2 < 0

where nLinhas = fromIntegral (length l)

nColunas = fromIntegral (length (head l))

Função auxiliar colisaoHitHit :

colisaoHitHit :: Hitbox -> Hitbox -> Bool

colisaoHitHit ((x1,y1),(x1',y1')) ((x2,y2),(x2',y2')) = if x1 < x2' && x1' > x2 && y1' < y2 && y1 > y2' then True else False

Função auxiliar colisaoHitboxparede:

colisaoHitboxparede :: Hitbox -> [Hitbox] -> Bool

colisaoHitboxparede hit [] = False

colisaoHitboxparede hit (h:t) = if colisaoHitHit hit h then True

else colisaoHitboxparede hit t

Exemplos de utiilização:

>>> colisaoHitHit ((3,6),(6,3)) ((5,4),(4,5))
True

>>> colisaoHitHit ((3,6),(6,3)) ((-1,-1),(-1,-1))
False

>>> criaHitbox (Personagem (5, -1) Jogador (0.5, 0.5) Oeste (1, 1) True True 2 0 (True, 5))
((4.5,-1.5),(5.5,-0.5))

colisoesPersonagens :: Personagem -> Personagem -> Bool

Função Principal.

criaHitbox :: Personagem -> Hitbox

Função auxiliar usada para obter a hitbox de um personagem em funcão da posição e tamanho.

hitboxMatriz :: [[Bloco]] -> [Hitbox]

Funções auxiliares usadas para obter as hitboxes dos blocos Plataforma e Alcapao de uma matriz de blocos.

hitboxLinha :: (Int, [Bloco]) -> [Hitbox]

Função auxiliar da função hitboxMatriz.

hitboxBloco :: Int -> Int -> Bloco -> [Hitbox]

Função auxiliar da função hitboxMatriz.

testePassalimite :: [[Bloco]] -> Hitbox -> Bool

Função que testa se uma hitbox passa dos limites do mapa.

colisaoHitHit :: Hitbox -> Hitbox -> Bool

Função auxiliar usada para verificar se duas hitboxes colidem.

colisaoHitboxparede :: Hitbox -> [Hitbox] -> Bool

Função recursiva da função colisaoHitHit que verifica se um personagem colide com os limites do mapa ou outra estrutura(plataformas,alcapões).