Tarefa1

Description

Módulo para a realização da Tarefa 1 de LI1 em 2023/24.

Documentation

tamanhoBlocos :: (Double, Double)

Representa o tamanho padrão dos blocos em termos de largura e altura na matriz. Isto é, todos os blocos representam 1 valor na matriz tanto na componente x, quanto na y.

hitbox :: Personagem -> Hitbox

Calcula a Hitbox de uma personagem. A hitbox é o menor retangulo que envolve uma personagem ou objeto. Nesta função, é dado o canto inferior esquerdo e o canto superior direito da hitbox. Os valores das posições dos limites da hitbox são dados através da posicao e tamanho do personagem.

Exemplo de utilização:

>>> hitbox (Personagem _ _ (1.0,1.0) _ (0.8,0.8) _ _ _ _ _)
((0.6,1.4),(1.4,0.6))

hitboxBloco :: Posicao -> Hitbox

Calcula a Hitbox de um bloco na matriz do mapa com base na posição do bloco na matriz e no tamanho padrão dos blocos. A hitbox é o menor retângulo que envolve uma personagem ou objeto. Nesta função é dado o canto inferior esquerdo e o canto superior direito da hitbox.

Exemplo de utilização:

>>> hitboxBloco (2.0,3.0)
((2.0,4.0),(3.0,3.0))

colisoesHitbox :: Hitbox -> Hitbox -> Bool

Verifica se duas Hitboxes colidem. Retorna True se as Hitboxes colidirem, False caso contrário.

Exemplo de utilização:

>>> colisoesHitbox ((1,1),(2,2)) ((1.5,1.5),(2.5,2.5))
False

colisaoLaterais :: Mapa -> Personagem -> Bool

Verifica se a personagem colide com as laterais do mapa através da posicao e do tamanho do mapa. Retorna True se houver colisão com as laterais.

Exemplo de utilização:

>>> colisaoLaterais m j
False

blocosPosicaoMatriz :: Mapa -> [(Bloco, Posicao)]

Obtém a lista de blocos na matriz do mapa com suas posições. Retorna uma lista de tuplos contendo o bloco e a posição na matriz.

Exemplo de utilização:

>>> blocosPosicaoMatriz (Mapa ((0.5, 0.5), Oeste) (0.5, 1.5) [[Plataforma, Plataforma, Vazio], [Plataforma, Plataforma, Vazio], [Plataforma, Plataforma, Vazio]])
[(Plataforma,(0.0,0.0)),(Plataforma,(1.0,0.0)),(Vazio,(2.0,0.0)),(Plataforma,(0.0,1.0)),(Plataforma,(1.0,1.0)),(Vazio,(2.0,1.0)),(Plataforma,(0.0,2.0)),(Plataforma,(1.0,2.0)),(Vazio,(2.0,2.0))]

blocosColididos :: [(Bloco, Posicao)] -> Personagem -> [(Bloco, Posicao)]

Obtém a lista de blocos na matriz do mapa que colidem com o personagem. Retorna uma lista de tuplos contendo o bloco e a posição na matriz que colidem com a Hitbox da personagem.

Exemplo de utilização:

>>> blocosColididos (blocosPosicaoMatriz m) j
[(Vazio,(2.0,1.0)),(Vazio,(3.0,1.0)),(Vazio,(2.0,2.0)),(Vazio,(3.0,2.0))]

colisaoPlataforma :: Mapa -> Personagem -> Bool

Verifica se a personagem colide com uma plataforma no mapa. Retorna True se houver colisão com uma plataforma.

Exemplo de utilização:

>>> colisaoPlataforma m j
False

colisoesParede :: Mapa -> Personagem -> Bool

Verifica se a personagem colide com as laterais do mapa ou com uma plataforma. Retorna True se houver colisão, False caso contrário.

Exemplo de utilização:

>>> colisoesParede m j
False

colisoesPersonagens :: Personagem -> Personagem -> Bool

Verifica se duas personagens colidem. Retorna True se as personagens colidirem.

Exemplo de utilização:

colisoesPersonagens j j1 True

colisaoParedeFrente :: Mapa -> Personagem -> Bool

Verifica se duas personagens colidem. Retorna True se as personagens colidirem.

Exemplo de utilização:

colisoesPersonagens j j1 True

colisaoLateraisDir :: Mapa -> Personagem -> Bool

Verifica se o personagem colidiu com as laterais direita do mapa.

Exemplo de utilização:

>>> colisaoLateraisDir m j
False

-