Tarefa2

Description

Módulo para a realização da Tarefa 2 de LI1 em 2023/24.

Documentation

valida :: Jogo -> Bool

Nesta tarefa, temos como objetivo definir a função valida, que tem como objetivo verificar se um dado jogo viola certas restrições previamente dadas:

O mapa tem “chão”, i.e. uma plataforma que impede que o jogador ou outro personagem caia fora do mapa.

Todos os inimigos tẽm a propriedade ressalta a True, enquanto que o jogador a tem a False.

A posição inicial de um jogador não pode colidir com a posição inicial de um outro personagem. Note que as posições iniciais de inimigos podem colidir entre estes.

Número mínimo de inimigos igual a 2 (dois).

inimigos Fantasma têm exatamente 1 (uma) vida.

Escadas não podem começar/terminar em alçaapões, e pelo menos uma das suas extremidades tem que ser do tipo Plataforma.

Alçapões não podem ser menos largos que o jogador.

Não podem existir personagens nem colecionãveis “dentro” de plataformas ou alçapões, isto é. o bloco (na matriz do mapa) correspendente à posição de um personagem ou objeto tem que ser Vazio.

Para a definir, precisamos usar as seguintes funções auxiliares:

minIni: testa se o número mínimo de inimigos no mapa é 2.

colidePI: testa se a posição inicial do jogador coincide com a posição inicial de outro personagem.

ress: testa se todos os inimigos têm a propriedade ressalta a True e o jogador a False.

existChao: testa se um mapa possui "chão", ou seja, uma plataforma que impede os personagens de cair fora do mapa.

tamanhoAlcp: testa se um jogador é menor que um bloco do tipo Alcapao.

fantLife: testa se todos os inimigo do tipo Fantasma tem uma vida.

colideVazio: testa se o bloco em que os personagens ou colecionáveis estão é do tipo Vazio ou Escada.

testEscd: testa se todas as escadas do mapa não começam nem terminam em alçapões, e se pelo menos uma das suas extremidades é do tipo Plataforma.

Funções:

Função Principal:

valida :: Jogo -> Bool

valida (Jogo mapa ini colc jog) = let cond1 = minIni ini

cond2 = not (colidePI mapa ini)

cond3 = ress ini jog

cond4 = existChao mapa

cond5 = tamanhoAlcp jog

cond6 = allTrue (fantLife ini)

cond7 = colideVazio ini jog colc mapa

cond8 = testEscd mapa (blcCimaBaixo mapa (cSupEsqE mapa))

in all (==True) [cond1,cond2,cond3,cond4,cond5,cond6,cond7,cond8]

Função auxiliar minIni:

minIni :: [Personagem] -> Bool

minIni ini = length ini >= 2

Função auxiliar colidePI e colAux:

colidePI :: Mapa -> [Personagem] -> Bool

colidePI (Mapa (posI,_) _ _) ini = any (== posI) (map colAux ini)

colAux :: Personagem -> Posicao

colAux indiv = (posicao indiv)

Função auxiliar ress e ressAux:

ress :: [Personagem] -> Personagem -> Bool

ress ini jog = all (== True) (map ressAux ini) && (ressalta jog) == False

ressAux :: Personagem -> Bool

ressAux indiv = (ressalta indiv)

Função auxiliar existChao:

existChao :: Mapa -> Bool

existChao (Mapa _ _ blc) = elem Plataforma (last blc)

Função auxiliar tamanhoAlcp:

tamanhoAlcp :: Personagem -> Bool

tamanhoAlcp jog = l <= 1 && c <= 1

where (c,l) = tamanho jog

Função auxiliar fantLife:

fantLife :: [Personagem] -> [Bool]

fantLife [] = []

fantLife (indiv:xs) = if (tipo indiv) == MacacoMalvado || (tipo indiv) == Jogador

then True : fantLife xs

else if (tipo indiv) == Fantasma && (vida indiv) == 1

then True : fantLife xs

else False : fantLife xs

Função auxiliar colideVazio:

colideVazio :: [Personagem] -> Personagem -> [(Colecionavel, Posicao)] -> Mapa -> Bool

colideVazio ini jog colc (Mapa _ _ blc) = all (`elem` (map posCentral ((cSupEsqV blc) ++ (cSupEsqEAux blc)))) (posElementos ini jog colc)

Função auxiliar posElementos:

posElementos :: [Personagem] -> Personagem -> [(Colecionavel, Posicao)] -> [Posicao]

posElementos ini jog colc =  map (\indiv-> (posicao indiv)) ini ++ [(posicao jog)] ++ map snd colc

Função auxiliar testEscd:

testEscd :: Mapa -> [(Bloco, Bloco)] -> Bool

testEscd (Mapa _ _ blc) list = if null (cSupEsqEAux blc)

then True

else if not (null (cSupEsqEAux blc)) && length list >= 1

then all (==True) (auxTestE list)

else False

Função auxiliar auxTesteE:

auxTestE :: [(Bloco,Bloco)] -> [Bool]

auxTestE [] = [True]

auxTestE ((x,y):r) = if x == Plataforma && (y == Plataforma || y == Vazio || y == Escada)

then True : auxTestE r

else if (x == Plataforma || x == Vazio || x == Escada) && y == Plataforma

then True : auxTestE r

else False : auxTestE r

Função auxiliar blcCimaBaixo:

blcCimaBaixo :: Mapa -> [Posicao] -> [(Bloco, Bloco)]

blcCimaBaixo _ [] = []

blcCimaBaixo (Mapa a b blc) ((x1, x2):xs) = (blcCima, blcBaixo) : blcCimaBaixo (Mapa a b blc) xs

where blcCima = if x1 > 0 then blc !! floor (x1 - 1) !! floor x2 else Vazio

blcBaixo = if x1 < fromIntegral (length blc) - 1 then blc !! floor (x1 + 1) !! floor x2 else Vazio

allTrue :: [Bool] -> Bool

Função testa que verifica se todos os valores lógicos de uma lista são do tipo True.

cSupEsqE :: Mapa -> [Posicao]

Função que dado um mapa dá uma lista das coordenadas do canto superior esquerdo de cada bloco do tipo Escada.

cSupEsqEAux :: [[Bloco]] -> [Posicao]

Função auxiliar de cSupEsqE.

cSupEsqV :: [[Bloco]] -> [Posicao]

Função que dado uma matriz de blocos dá uma lista das coordenadas do canto superior esquerdo de cada bloco do tipo Vazio.

cSupEsqP :: [[Bloco]] -> [Posicao]

Função que dado uma matriz de blocos dá uma lista das coordenadas do canto superior esquerdo de cada bloco do tipo Plataforma.

posCentral :: Posicao -> Posicao

Função que retorna as coodenadas da posição central de um bloco apartir das coordenadas do canto superior esquerdo desse mesmo bloco.

minIni :: [Personagem] -> Bool

Função que testa se o número mínimo de inimigos no mapa é 2.

colidePI :: Mapa -> [Personagem] -> Bool

Função que testa se a posição inicial do jogador coincide com a posição inicial de outro personagem.

colAux :: Personagem -> Posicao

Função auxiliar de colidePi.

ress :: [Personagem] -> Personagem -> Bool

Função que testa se todos os inimigos têm a propriedade ressalta a True e o jogador a False.

ressAux :: Personagem -> Bool

Função auxiliar de ress.

existChao :: Mapa -> Bool

Função que testa se um mapa possui "chão".

tamanhoAlcp :: Personagem -> Bool

Função que testa se um jogador é menor que um bloco do tipo Alcapao.

fantLife :: [Personagem] -> [Bool]

Função que testa se todos os inimigo do tipo Fantasma tem uma vida.

colideVazio :: [Personagem] -> Personagem -> [(Colecionavel, Posicao)] -> Mapa -> Bool

Função que testa se o bloco em que os personagens ou colecionáveis estão é do tipo Vazio ou Escada.

posElementos :: [Personagem] -> Personagem -> [(Colecionavel, Posicao)] -> [Posicao]

Função que calcula a lista de posições de todos os personagens.

testEscd :: Mapa -> [(Bloco, Bloco)] -> Bool

Função que testa se todas as escadas do mapa não começam nem terminam em alçapões, e se pelo menos uma das suas extremidades é do tipo Plataforma.

auxTestE :: [(Bloco, Bloco)] -> [Bool]

Função auxiliar que recebe uma lista de pares dos blocos acima e em baixo das escadas e devole a lista dos valores lógicos se estas apresentarem pelo menos um bloco de plataforma e nenhum bloco de alçapão.

blcCimaBaixo :: Mapa -> [Posicao] -> [(Bloco, Bloco)]

Função que dada uma matriz de blocos e uma lista das coordenadas do canto superior esquerdo de cada bloco do tipo Escada dessa matriz(calculada pela findblocE), devolve uma lista de pares que correspondem ao bloco acima e em baixo das escada, respetivamente.