Tarefa2

Description

Módulo para a realização da Tarefa 2 de LI1 em 2023/24.

Documentation

valida :: Jogo -> Bool

Função principal que valida o estado do jogo. Verifica várias condições, incluindo validação do chão, ressalto, posição, inimigos, vida, escadas, alçapões e blocos. Retorna True se todas as verificações passarem, indicando um estado de jogo válido.

validaChao :: Mapa -> Bool

Verifica se a última linha do mapa é composta exclusivamente por plataformas. Retorna True se a última linha contiver apenas plataformas, indicando um chão sólido.

Exemplos de utilização:

>>> validaChao (Mapa (5, 3) [Plataforma] [[Plataforma, Plataforma, Plataforma], [Plataforma, Plataforma, Plataforma], [Plataforma, Plataforma, Plataforma]])
True
>>> validaChao (Mapa (5, 3) [Plataforma] [[Plataforma, Plataforma, Plataforma], [Vazio, Vazio, Vazio], [Plataforma, Plataforma, Plataforma]])
False 

validaRessalto :: Jogo -> Bool

validaPosicao :: Jogo -> Bool

Verifica se a posição do jogador é diferente da posição de todos os inimigos. Retorna True se a posição do jogador não colidir com nenhum inimigo.

Exemplos de utilização:

>>> validaPosicao j
True

validaInimigos :: Jogo -> Bool

Verifica se há pelo menos dois inimigos no jogo. Retorna True se houver pelo menos dois inimigos no jogo.

Exemplos de utilização:

>>> validaInimigos (Jogo m [] [(Martelo,(5, 3)), (Moeda, (8, 2))]  j)
False

>>> validaInimigos (Jogo m [Fantasma, Fantasma] [(Martelo,(5, 3)), (Moeda, (8, 2))] j)
True 

validaVida :: [Personagem] -> Bool

Verifica se os fantasmas têm apenas uma vida. Retorna True se todos os fantasmas têm exatamente uma vida.

Exemplos de utilização:

>>> validaVida en1
True

>>> validaVida en3
False

verificarColuna :: [Bloco] -> Bool

Verifica as condições em cada coluna do mapa, fazendo esta verificação em linha no mapa transposto. Condições: as escadas começam ou acabam com uma plataforma e nunca com um alçapão.

Exemplos de utilização:

>>> verificarColuna [Plataforma, Plataforma, Escada, Plataforma]
True

>>> verificarColuna [Plataforma, Escada, Alcapao, Plataforma]
False 

validaEscadas :: Mapa -> Bool

Usa a função verificarColuna no mapa transposto (usando a função transpose) para verificar se as escadas estão dentro das condições estabelecidas.

Exemplos de utilização:

>>> validaEscadas m
True

>>> validaEscadas (Mapa ((0.5, 5.5), Oeste) (0.5, 2.5) [[Escada, Escada, Escada], [Vazio, Vazio, Vazio], [Plataforma, Plataforma, Plataforma]])
False 

validaAlcapao :: Jogo -> Bool

Verifica se a dimensão do alçapão é maior ou igual à largura do jogador. Retorna True se a largura do jogador couber dentro das dimensões do alçapão.

Exemplos de utilização: >>> validaAlcapao jg True

validaBlocos :: Jogo -> Bool

Verifica se a posição do jogador coincide com um bloco de plataforma ou alçapão. Retorna True se a posição do jogador não estiver sobre uma plataforma ou alçapão.

Exemplos de utilização:

>>> validaBlocos (Jogo (Mapa ((0.5, 5.5), Oeste) (0.5, 2.5) [[Vazio, Vazio, Plataforma], [Vazio, Vazio, Vazio], [Vazio, Vazio, Vazio]]) [en1, en2] [(Martelo,(5, 3)), (Moeda, (8, 2))] j)
True