Tarefa3

Description

Módulo para a realização da Tarefa 3 de LI1 em 2023/24.

Documentation

movimenta :: Semente -> Tempo -> Jogo -> Jogo

Atualiza o estado do jogo, movendo todos os elementos de acordo com o tempo. Retorna um novo estado de jogo após a movimentação.

Exemplos de utilização:

>>> movimenta 7 1.3 jg
Jogo {mapa = Mapa ((0.5,5.5),Oeste) (0.5,2.5) [[Plataforma,Plataforma,Plataforma,Plataforma,Plataforma,Plataforma,Plataforma,Plataforma,Plataforma,Plataforma],[Vazio,Vazio,Vazio,Vazio,Vazio,Vazio,Vazio,Vazio,Vazio,Vazio],[Vazio,Vazio,Vazio,Vazio,Vazio,Vazio,Vazio,Vazio,Vazio,Vazio],[Plataforma,Plataforma,Vazio,Vazio,Vazio,Vazio,Plataforma,Plataforma,Plataforma,Plataforma],[Vazio,Vazio,Vazio,Vazio,Vazio,Vazio,Vazio,Vazio,Escada,Vazio],[Vazio,Vazio,Vazio,Vazio,Vazio,Vazio,Vazio,Vazio,Escada,Vazio],[Plataforma,Plataforma,Plataforma,Plataforma,Plataforma,Plataforma,Plataforma,Plataforma,Plataforma,Plataforma]]
, inimigos = [Personagem {velocidade = (1.0,0.0), tipo = Fantasma, posicao = (1.0,2.0), direcao = Este, tamanho = (1.0,1.0), emEscada = False, ressalta = True, vida = 2, pontos = 0, aplicaDano = (True,1.0)},Personagem {velocidade = (1.0,0.0), tipo = Fantasma, posicao = (1.0,1.0), direcao = Este, tamanho = (1.0,1.0), emEscada = False, ressalta = True, vida = 2, pontos = 0, aplicaDano = (True,1.0)}], colecionaveis = [(Moeda,(5.0,3.0)),(Moeda,(8.0,2.0))]
, jogador = Personagem {velocidade = (0.0,10.0), tipo = Jogador, posicao = (3.0,2.0), direcao = Este, tamanho = (1.0,1.0), emEscada = False, ressalta = False, vida = 3, pontos = 0, aplicaDano = (False,0.0)}}

hitboxDano :: Personagem -> Hitbox

Calcula a hitbox de dano de uma personagem com base na direção. Retorna a hitbox de dano da personagem.

Exemplos de utilização:

>>> hitboxDano j1
((3.5,2.5),(4.5,1.5))

hitboxColecionavel :: (Colecionavel, Posicao) -> Hitbox

Calcula a hitbox de um colecionável com base em sua posição.

Exemplos de utilização: >>> hitboxColecionavel (Moeda, (3.5, 1.0)) ((3.0,1.5),(4.0,0.5))

>>> hitboxColecionavel (Martelo, (1.5, 2.0))
(1.0,2.5),(2.0,1.5))

deslocarInimigos :: [Personagem] -> Jogo -> Semente -> Tempo -> [Personagem]

Desloca os inimigos com base no tempo, semente e jogo. Retorna uma lista de inimigos com posições atualizadas.

Exemplo de utilização:

>>> deslocarInimigos i jg 3 2.0
[Personagem {velocidade = (1.0,0.0), tipo = Fantasma, posicao = (1.0,2.0), direcao = Oeste, tamanho = (1.0,1.0), emEscada = False, ressalta = True, vida = 2, pontos = 0, aplicaDano = (True,1.0)}
,Personagem {velocidade = (1.0,0.0), tipo = Fantasma, posicao = (1.0,1.0), direcao = Oeste, tamanho = (1.0,1.0), emEscada = False, ressalta = True, vida = 2, pontos = 0, aplicaDano = (True,1.0)}]  

direcaoContraria :: Direcao -> Direcao

Inverte a direção passada como argumento

Exemplo de utilização:

>>> direçaoContraria Norte
Sul

deslocarJogador :: Personagem -> Mapa -> Tempo -> Personagem

Desloca o jogador com base no tempo e mapa. Retorna o jogador com a posição atualizada.

Exemplo de utilização:

>>> deslocarJogador j m 7.9
Personagem {velocidade = (0.0,6.0), tipo = Jogador, posicao = (3.0,2.0), direcao = Sul, tamanho = (1.0,1.0), emEscada = False, ressalta = False, vida = 3, pontos = 0, aplicaDano = (False,0.0)}

colecionaveisJog :: [(Colecionavel, Posicao)] -> Personagem -> (Personagem, [(Colecionavel, Posicao)])

Verifica a interação entre o jogador e os colecionáveis. Retorna o jogador e a lista atualizada de colecionáveis.

Exemplo de utilização:

>>> colecionaveisJog [(Moeda,(5.0, 3.0)), (Moeda, (8.0, 2.0))]  j
(Personagem {velocidade = (0.0,0.0), tipo = Jogador, posicao = (3.0,2.0), direcao = Este, tamanho = (1.0,1.0), emEscada = False, ressalta = False, vida = 3, pontos = 0, aplicaDano = (False,0.0)},[(Moeda,(5.0,3.0)),(Moeda,(8.0,2.0))])

interacaoJogInim :: [Personagem] -> Personagem -> (Personagem, [Personagem])

Verifica a interação entre o jogador e os inimigos. Retorna o jogador e a lista atualizada de inimigos.

Exemplo de utilização 1:

>>> interacaoJogInim i j
(Personagem {velocidade = (0.0,0.0), tipo = Jogador, posicao = (3.0,2.0), direcao = Este, tamanho = (1.0,1.0), emEscada = False, ressalta = False, vida = 3, pontos = 0, aplicaDano = (False,0.0)}
,[Personagem {velocidade = (1.0,0.0), tipo = Fantasma, posicao = (1.0,2.0), direcao = Este, tamanho = (1.0,1.0), emEscada = False, ressalta = True, vida = 2, pontos = 0, aplicaDano = (True,1.0)}
,Personagem {velocidade = (1.0,0.0), tipo = Fantasma, posicao = (1.0,1.0), direcao = Este, tamanho = (1.0,1.0), emEscada = False, ressalta = True, vida = 2, pontos = 0, aplicaDano = (True,1.0)}])

danoInimigo :: [Personagem] -> Personagem -> [Personagem]

Aplica dano aos inimigos com base na hitbox de dano do jogador. Retorna a lista atualizada de inimigos após o dano.

Exemplo de utilização:

>>> danoInimigo i j
[Personagem {velocidade = (1.0,0.0), tipo = Fantasma, posicao = (1.0,2.0), direcao = Este, tamanho = (1.0,1.0), emEscada = False, ressalta = True, vida = 2, pontos = 0, aplicaDano = (True,1.0)}
,Personagem {velocidade = (1.0,0.0), tipo = Fantasma, posicao = (1.0,1.0), direcao = Este, tamanho = (1.0,1.0), emEscada = False, ressalta = True, vida = 2, pontos = 0, aplicaDano = (True,1.0)}]

checkAlcapao :: Mapa -> Personagem -> Mapa

Verifica a existência de um alçapão abaixo do jogador e altera o mapa se encontrado. Retorna um novo mapa após a verificação e possível alteração.

Exemplo de utilização:

>>> checkAlcapao m j
Mapa ((0.5,5.5),Oeste) (0.5,2.5) [[Plataforma,Plataforma,Plataforma,Plataforma,Plataforma,Plataforma,Plataforma,Plataforma,Plataforma,Plataforma]
                                 ,[Vazio,Vazio,Vazio,Vazio,Vazio,Vazio,Vazio,Vazio,Vazio,Vazio]
                                 ,[Vazio,Vazio,Vazio,Vazio,Vazio,Vazio,Vazio,Vazio,Vazio,Vazio]
                                 ,[Plataforma,Plataforma,Vazio,Vazio,Vazio,Vazio,Plataforma,Plataforma,Plataforma,Plataforma]
                                 ,[Vazio,Vazio,Vazio,Vazio,Vazio,Vazio,Vazio,Vazio,Escada,Vazio],[Vazio,Vazio,Vazio,Vazio,Vazio,Vazio,Vazio,Vazio,Escada,Vazio]
                                 ,[Plataforma,Plataforma,Plataforma,Plataforma,Plataforma,Plataforma,Plataforma,Plataforma,Plataforma,Plataforma]]

alteracaoMatriz :: [[Bloco]] -> Posicao -> [[Bloco]]

Altera uma matriz de blocos, marcando um alçapão como Vazio. Retorna uma matriz de blocos após a alteração.

- Exemplo de utilização: >>> alteracaoMatriz [[Plataforma, Plataforma, Plataforma, Plataforma, Plataforma],[Vazio, Alcapao, Vazio],[Vazio, Vazio, Vazio]] (1.5,1.5) [[Plataforma,Plataforma,Plataforma,Plataforma,Plataforma],[Vazio,Vazio,Vazio],[Vazio,Vazio,Vazio]]

calcDeslocamento :: Velocidade -> Tempo -> Posicao

Calcula o deslocamento com base na velocidade e tempo. Retorna a nova posição após o deslocamento.

Exemplos de utilização:

>>> calcDeslocamento (2, 1) 0.5
(1.0, 0.5)

>>> calcDeslocamento (0, -1) 1.0
(0.0, -1.0)

decisao :: Semente -> Bool

Decide com base na semente se um evento aleatório deve ocorrer. Retorna True se o evento deve ocorrer.

Exemplo de utilização:

>>> decisao 3
True

>>> decisao 47573
False

emPlataforma :: Mapa -> Personagem -> Bool

Verifica se a personagem está em uma plataforma no mapa.

Exemplo de utilização:

>>> emPlataforma m j
False

>>> emPlataforma m j2
True