Tarefa4

Description

Módulo para a realização da Tarefa 4 de LI1 em 2023/24.

Documentation

atualiza :: [Maybe Acao] -> Maybe Acao -> Jogo -> Jogo

Nesta tarefa, temos como objetivo definir a função atualiza que deve atualizar as novas direções e velocidades dos personagens de acordo com as ações dadas.

Para isso, precisamos de definir as seguintes funções auxiliares:

atuaInimigo, que trata de atualizar os inimigos.

atuaJogador, que trata de atualizar o Jogador.

Cada uma destas funções vai usar a função atuaAcao, que dado um mapa, um personagem individual e uma ação atualiza esse personagem.

Funções:

Função Principal:

atualiza :: [Maybe Acao] -> Maybe Acao -> Jogo -> Jogo

atualiza acIni acJog jogo = jogo { inimigos = novoIni, jogador = novoJog }

where novoIni = atuaInimigo (mapa jogo) acIni (inimigos jogo)

novoJog = atuaJogador (mapa jogo) acJog (jogador jogo)

Função auxiliar atuaInimigo:

atuaInimigo :: Mapa -> [Maybe Acao] -> [Personagem] -> [Personagem]

atuaInimigo _ [] ini = ini

atuaInimigo mapa (h:t) (x:xs) = case h of

Just a -> atuaAcao mapa a x : atuaInimigo mapa t xs

Nothing -> (x:xs)

Função auxiliar atuaJogador:

atuaJogador :: Mapa -> Maybe Acao -> Personagem -> Personagem

atuaJogador mapa ac jog = case ac of

Just a -> atuaAcao mapa a jog

Nothing -> jog

Função auxiliar atuaAcao:

atuaAcao :: Mapa -> Acao -> Personagem -> Personagem

atuaAcao (Mapa _ _ blc) ac pers = case ac of

Subir -> if sobEsc blc pers then pers { velocidade = (0,-5), direcao = Norte, emEscada = True} else pers

Descer -> if sobEsc blc pers then pers { velocidade = (0,5), direcao = Sul, emEscada = True} else pers

AndarDireita -> if movDir blc pers then pers { velocidade = (-5,0), direcao = Oeste} else pers { velocidade = (5,0), direcao = Este}

AndarEsquerda -> if movEsq blc pers then pers { velocidade = (5,0), direcao = Este} else pers { velocidade = (-5,0), direcao = Oeste}

Saltar -> if (emEscada pers) == False then pers { velocidade = (0,-5)} else pers

Parar -> if permParar blc pers == True then pers { velocidade = (0,0)} else pers

Exemplos de utiilização:

>>> atualiza ([Nothing,Hothing]) (Nothing) Jogo (Mapa ((4.5, 2), Oeste) (5, 5) [[Plataforma, Plataforma, Plataforma, Alcapao, Alcapao], [Vazio, Plataforma, Escada, Plataforma, Vazio], [Vazio, Vazio, Vazio, Vazio, Vazio], [Vazio, Escada, Vazio, Vazio, Escada], [Plataforma, Plataforma, Alcapao, Plataforma, Plataforma]]) ([(Personagem (4, 0) Fantasma (4.5, 1.5) Oeste (1, 1) False True 1 0 (False, 0)),(Personagem (1, 5) Fantasma (0.5, 1.5) Este (1, 1) False True 1 0 (False, 0)),(Personagem (1, 0) Fantasma (2.5, 2.5) Oeste (1, 1) False True 0 0 (False, 0))]) ([(Martelo,(4.5,4.5)),(Moeda,(1.5,1.5))]) (Personagem (0.5, 0.5) Jogador (0.5, 1.5) Norte (1, 1) False True 2 0 (True, 10))
Jogo (Mapa ((4.5, 2), Oeste) (5, 5) [[Plataforma, Plataforma, Plataforma, Alcapao, Alcapao], [Vazio, Plataforma, Escada, Plataforma, Vazio], [Vazio, Vazio, Vazio, Vazio, Vazio], [Vazio, Escada, Vazio, Vazio, Escada], [Plataforma, Plataforma, Alcapao, Plataforma, Plataforma]]) ([(Personagem (4, 0) Fantasma (4.5, 1.5) Oeste (1, 1) False True 1 0 (False, 0)),(Personagem (1, 5) Fantasma (0.5, 1.5) Este (1, 1) False True 1 0 (False, 0)),(Personagem (1, 0) Fantasma (2.5, 2.5) Oeste (1, 1) False True 0 0 (False, 0))]) ([(Martelo,(4.5,4.5)),(Moeda,(1.5,1.5))]) (Personagem (0.5, 0.5) Jogador (0.5, 1.5) Norte (1, 1) False True 2 0 (True, 10))

hitBlocosCEsq :: [Posicao] -> [Hitbox]

Função que calcula a hitbox de uma lista de blocos dado a lista das coordenadas do canto superior esquerdo.

atuaInimigo :: Mapa -> [Maybe Acao] -> [Personagem] -> [Personagem]

Função que atualiza uma lista de inimigos de acordo com a lista de ações dadas.

atuaJogador :: Mapa -> Maybe Acao -> Personagem -> Personagem

Função que atualiza a personagem do jogador de acordo com a ação dada.

atuaAcao :: Mapa -> Acao -> Personagem -> Personagem

Função que, dado um mapa, uma ação e um personagem atualiza o respetivo personagem.

sobEsc :: [[Bloco]] -> Personagem -> Bool

Função que verifica se uma personagem se encontra na posição de um bloco do tipo Escada.

movDir :: [[Bloco]] -> Personagem -> Bool

Função que verifica se uma personagem não se pode mover para a direita.

movEsq :: [[Bloco]] -> Personagem -> Bool

Função que testa se uma personagem não se pode mover para a esquerda.

permParar :: [[Bloco]] -> Personagem -> Bool

Função que testa se um personagem pode estar parado.