Tarefa4

Description

Módulo para a realização da Tarefa 4 de LI1 em 2023/24.

Documentation

atualiza :: [Maybe Acao] -> Maybe Acao -> Jogo -> Jogo

Atualiza o jogo com base nas ações do jogador e inimigos. Retorna o novo jogo após aplicar as atualizações.

Exemplo de utilização:

>>> atualiza [(Just Subir), (Just AndarDireita)] (Just Descer) jg
Jogo {mapa = Mapa ((0.5,5.5),Oeste) (0.5,2.5) [[Plataforma,Plataforma,Plataforma,Plataforma,Plataforma,Plataforma,Plataforma,Plataforma,Plataforma,Plataforma]
,[Vazio,Vazio,Vazio,Vazio,Vazio,Vazio,Vazio,Vazio,Vazio,Vazio],[Vazio,Vazio,Vazio,Vazio,Vazio,Vazio,Vazio,Vazio,Vazio,Vazio]
,[Plataforma,Plataforma,Vazio,Vazio,Vazio,Vazio,Plataforma,Plataforma,Plataforma,Plataforma],[Vazio,Vazio,Vazio,Vazio,Vazio,Vazio,Vazio,Vazio,Escada,Vazio]
,[Vazio,Vazio,Vazio,Vazio,Vazio,Vazio,Vazio,Vazio,Escada,Vazio],[Plataforma,Plataforma,Plataforma,Plataforma,Plataforma,Plataforma,Plataforma,Plataforma,Plataforma,Plataforma]]
, inimigos = [Personagem {velocidade = (0.0,0.0), tipo = Fantasma, posicao = (1.0,2.0), direcao = Este, tamanho = (1.0,1.0), emEscada = False, ressalta = True, vida = 2, pontos = 0, aplicaDano = (True,1.0)}
,Personagem {velocidade = (0.0,6.0), tipo = Fantasma, posicao = (1.0,1.0), direcao = Sul, tamanho = (1.0,1.0), emEscada = False, ressalta = True, vida = 2, pontos = 0, aplicaDano = (True,1.0)}], colecionaveis = [(Moeda,(5.0,3.0)),(Moeda,(8.0,2.0))]
, jogador = Personagem {velocidade = (0.0,0.0), tipo = Jogador, posicao = (3.0,2.0), direcao = Este, tamanho = (1.0,1.0), emEscada = False, ressalta = False, vida = 3, pontos = 0, aplicaDano = (False,0.0)}}

atualizaInimigos :: [(Maybe Acao, Personagem)] -> Jogo -> [Personagem]

Atualiza a lista de inimigos com base nas ações e no estado atual do jogo. Retorna a lista de inimigos atualizada.

Exemplo de utilização:

>>> atualizaInimigos [(Just Subir, en1), (Just AndarDireita, en2)] jg
[Personagem {velocidade = (0.0,0.0), tipo = Fantasma, posicao = (4.5,10.5), direcao = Este, tamanho = (1.0,1.0), emEscada = False, ressalta = True, vida = 1, pontos = 0, aplicaDano = (False,0.0)}
,Personagem {velocidade = (0.0,6.0), tipo = MacacoMalvado, posicao = (4.5,1.5), direcao = Sul, tamanho = (1.0,1.0), emEscada = False, ressalta = True, vida = 10, pontos = 0, aplicaDano = (False,0.0)}]

atualizaAcao :: Maybe Acao -> Jogo -> Personagem -> Personagem

Atualiza as ações do jogador e inimigos com base no estado atual do jogo. Retorna a personagem atualizada com base na ação recebida.

Exemplo de utilização:

>>> atualizaAcao (Just AndarDireita) jg j2
Personagem {velocidade = (1.0,0.0), tipo = Jogador, posicao = (1.5,2.5), direcao = Este, tamanho = (1.0,1.0), emEscada = False, ressalta = False, vida = 3, pontos = 0, aplicaDano = (False,0.0)}

escada :: Personagem -> Mapa -> Bool

Verifica se a personagem está em um bloco escada.

Exemplos de utiliação:

>>> escada j m
False

up :: Jogo -> Personagem -> Personagem

Move a personagem para cima, verificando se está em uma escada (ou numa plataforma, em certos casos). Retorna a personagem atualizada com nova posição ou estado em escada.

Exemplos de utilização:

>>> up jg j
Personagem {velocidade = (0.0,0.0), tipo = Jogador, posicao = (3.0,2.0), direcao = Este, tamanho = (1.0,1.0), emEscada = False, ressalta = False, vida = 3, pontos = 0, aplicaDano = (False,0.0)}

> 

blocoAbaixo :: Posicao -> Mapa -> Maybe Bloco

procuraBloco :: Posicao -> [(Bloco, Posicao)] -> Maybe Bloco

Dada uma posição, busca um bloco em uma lista de tuplas (bloco, posição).

Exemplo de utilização

>>> procuraBloco (2.0,3.0) [(Escada, (2.0,3.0)), (Plataforma, (3.0,4.0)), (Vazio, (1.0,9.0))]
Just Escada

down :: Jogo -> Personagem -> Personagem

Move a personagem para baixo, verificando se está em uma escada. Retorna a personagem atualizada com nova posição ou estado em escada.

Exemplos de utilização:

>>> down jg j
Personagem {velocidade = (0.0,0.0), tipo = Jogador, posicao = (3.0,2.0), direcao = Este, tamanho = (1.0,1.0), emEscada = False, ressalta = False, vida = 3, pontos = 0, aplicaDano = (False,0.0)}

andarD :: Jogo -> Personagem -> Personagem

Move a personagem para a direita, verificando colisões e ressalto. Retorna a personagem atualizada com nova posição ou estado de ressalto.

Exemplos de utilização:

>>> andarD jg j2
Personagem {velocidade = (1.0,0.0), tipo = Jogador, posicao = (1.5,2.5), direcao = Este, tamanho = (1.0,1.0), emEscada = False, ressalta = False, vida = 3, pontos = 0, aplicaDano = (False,0.0)}

andarE :: Jogo -> Personagem -> Personagem

Move a personagem para a esquerda, verificando colisões e ressalto. Retorna a personagem atualizada com nova posição ou estado de ressalto.

Exemplos de utilização:

>>> andarE jg j
Personagem {velocidade = (-1.0,0.0), tipo = Jogador, posicao = (3.0,2.0), direcao = Este, tamanho = (1.0,1.0), emEscada = False, ressalta = False, vida = 3, pontos = 0, aplicaDano = (False,0.0)}

verificaColisaoInimigos :: Jogo -> Personagem -> Bool

Verifica colisões de um certo inimigo com os outros inimigos. Retorna True se a personagem colidir com algum inimigo.

Exemplos de utilização:

>>> verificaColisaoInimigos jg j
False

queda :: Personagem -> Mapa -> Bool

Verifica se a personagem está em queda livre. Retorna True se a personagem não está em uma escada e está em queda.

Exemplos de utilização:

>>> queda j m
False

verificaP :: Mapa -> Personagem -> Bool

Verifica se os personagens podem cair, ou seja, se o bloco imediatamente a seguir do bloco abaixo da personagem é vazio Não pode estar em uma escada e não há colisão com Plataforma ou Alçapão.

Exemplos de utilização:

>>> verificaP m j
False

saltar :: Mapa -> Personagem -> Personagem

Faz a personagem realizar um salto. Retorna a personagem com velocidade de salto aplicada.

Exemplos de utilização: >>> saltar m j2 Personagem {velocidade = (0.0,-40.0), tipo = Jogador, posicao = (1.5,2.5), direcao = Norte, tamanho = (1.0,1.0), emEscada = False, ressalta = False, vida = 3, pontos = 0, aplicaDano = (False,0.0)}

>>> saltar m j
Personagem {velocidade = (0.0,6.0), tipo = Jogador, posicao = (3.0,2.0), direcao = Sul, tamanho = (1.0,1.0), emEscada = False, ressalta = False, vida = 3, pontos = 0, aplicaDano = (False,0.0)}

parar :: Personagem -> Mapa -> Personagem

Faz a personagem parar. Retorna a personagem com velocidade igual a 0, exceto se estiver em queda.

Exemplos de utilização:

>>> parar j m
Personagem {velocidade = (0.0,0.0), tipo = Jogador, posicao = (3.0,2.0), direcao = Este, tamanho = (1.0,1.0), emEscada = False, ressalta = False, vida = 3, pontos = 0, aplicaDano = (False,0.0)}