Tarefa5

Description

Módulo para a realização da Tarefa 5 de LI1 em 2023/24.

Documentation

main :: IO ()

Nesta parte do trabalho temos como objetivo usar as funcionalidades e funções já elaboradas em outras Tarefas, juntamente da biblioteca Gloss para construir uma aplicação gráfica que permita o jogador jogar.

Para isso, usamos a função play da biblioteca Graphics.Gloss.Interface.Pure.Game e definimos:

Um novo estado de jogo, chamado GameState, que vai representar todo o estado do jogo, sendo do tipo Menu TipoMenu Jogo ou InGame Jogo;

Uma janela, do tipo Display, que nos vai permitir visualizar o jogo;

A cor de fundo que a nossa janela vai tomar;

O frame rate, ou seja, o número de frames por segundo que a janela vai mostrar;

Um estado inicial, que vai ser do tipo Menu Inicial jogoInicial1;

Uma função responsável por determinar como o estado do jogo será visualizado com gráficos 2D, ou seja, como se converte para um valor do tipo Picture.

A essa função demos o nome animar que converte GameState em Picture.

Para a definir usamos duas funções auxiliares:

animarMenu, que é responsável por animar o diferentes tipos de menu que existem, sendo eles Inicial, Pausa, Fim e Comandos;

animarJogo, que é responsável por animar o mapa, os inimigos, os colecionaveis e o jogador , que pertencem ao Jogo.

Uma função responsável por atualizar o GameState dependendo das ações do jogador, ou seja, dos possíveis eventos;

Demos-lhe o nome reageEvento que atualiza o GameState.

A função é responsável pelas ações do jogador tanto no estado InGame Jogo, quanto no estado Menu TipoMenu Jogo.

Além disso, usamos a função atualiza da Tarefa4 que dadas certas ações do jogador no estado InGame Jogo gera ações para o jogador e inimigos.

Uma função responsável por atualizar o GameState em consequência da passagem do tempo.

Demos-lhe o nome reageTempo, na qual implementamos a função movimenta da Tarefa3 que atualiza as posições dos personagens e trata de possíveis consequências.

Argumentos da função play:

Função Main:

main :: IO ()

main = do play dm white fr estadoInit animar reageEvento reageTempo

Janela

dm :: Display

dm = InWindow "Donkey Kong" (700, 700) (0, 0)

Frame Rate:

fr :: Int

fr = 60

Estado Inicial:

estadoInit :: GameState

estadoInit = Menu Inicial jogoInicial1

Função animar:

animar :: GameState -> Picture

animar (Menu tipoMenu jogo) = animarMenu tipoMenu jogo

animar (InGame jogo) = animarJogo jogo

Função reageEvento:

reageEvento :: Event -> GameState -> GameState

reageEvento evento (Menu h jogo) = eventosMenu evento (Menu h jogo)

reageEvento evento (InGame jogo) = eventosEmJogoJogador evento (InGame novoJogo2)

where (InGame novoJogo2) = InGame (eventosEmJogoInimigos (jogo))

Função reageTempo:

reageTempo :: Float -> GameState -> GameState

reageTempo t (Menu tipo jogo) = Menu tipo jogo

reageTempo t (InGame jogo) = InGame (movimenta 20 (realToFrac t) jogo)

dm :: Display

Configurações da janela de visualização.

fr :: Int

Números de frames por segundo.

estadoInit :: GameState

Estado inicial do jogo.

data TipoMenu

Tipos de Menu

Constructors

Inicial

Pausa

Fim

Comandos

data GameState

Tipos de estado do jogo.

Constructors

Menu TipoMenu Jogo

InGame Jogo

jogoInicial1 :: Jogo

Jogo Inicial.

mapaInicial1 :: Mapa

Mapa Inicial.

inimigosInicial1 :: [Personagem]

Inimigos do Jogo Inicial.

colcInicial1 :: [(Colecionavel, (Double, Double))]

Colecionáveis do Jogo Inicial.

jogadorInicial1 :: Personagem

Jogador Inicial.

fantasma1 :: Personagem

fantasma2 :: Personagem

plataforma :: Picture

Defenição do bloco Plataforma em Picture.

alcapao :: Picture

Defenição do bloco Alcapao em Picture.

escada :: Picture

Defenição do bloco Escada em Picture.

vazio :: Picture

Defenição do bloco Vazio em Picture.

blocoToPic :: Bloco -> Picture

Função auxiliar que passa de Bloco para Picture.

moeda :: Picture

Defenição do colecionável Moeda em Picture.

martelo :: Picture

Defenição do colecionável Martelo em Picture.

colcToPic :: Colecionavel -> Picture

Função auxiliar que passa de Colecionavel para Picture.

animar :: GameState -> Picture

Função para animar o estado do jogo.

animarMenu :: TipoMenu -> Jogo -> Picture

Função para animar o menu.

animarMenuInicial :: TipoMenu -> Picture

Função auxiliar para animar o menu inicial.

animarMenuPausa :: TipoMenu -> Picture

Função auxiliar para animar o menu pausa.

animarMenuFim :: TipoMenu -> Jogo -> Picture

Função auxiliar para animar o menu final.

animarMenuComandos :: TipoMenu -> Picture

Função auxiliar para animar o menu contendo os comandos.

animarJogo :: Jogo -> Picture

Função para animar o jogo.

desenharLinha :: Int -> [Bloco] -> [Picture]

Função para animar uma linha da matriz que representa o mapa.

animarMapa :: (Int, Int) -> [[Bloco]] -> [Picture]

Função para animar a matriz que representa o mapa.

translateInimigos :: [Personagem] -> [Picture]

Função para animar os inimigos do Jogo.

translateColc :: [(Colecionavel, Posicao)] -> [Picture]

Função para animar os colecionaveis do Jogo.

reageEvento :: Event -> GameState -> GameState

Função para lidar com as ações do jogador.

eventosEmJogoInimigos :: Jogo -> Jogo

Função que adiciona ações aleatórias aos Inimigos.

criaAcoes :: StdGen -> Int -> [Maybe Acao]

Função auxiliar de eventosEmJogoInimigos.

e :: Int

d :: Int

criaAcaoIndiv :: StdGen -> Maybe Acao

Função auxiliar de eventosEmJogoInimigos.

eventosMenu :: Event -> GameState -> GameState

Função para lidar com as ações do jogador quando este está em algum Menu.

eventosEmJogoJogador :: Event -> GameState -> GameState

Função para lidar com as ações do jogador quando este está em algum Menu.

reageTempo :: Float -> GameState -> GameState

Função que calcula as novas posições dos personagens e respetivas consequências conforme passa o tempo.