Tarefa5

Description

Módulo para a realização da Tarefa 5 de LI1 em 2023/24.

Documentation

data Estado

Constructors

Estado

Fields

jogoAtual :: Jogo

imagens :: Imagens

modo :: Modo

jogoClean :: Jogo

acaoJog :: Maybe Acao

menu :: MenuInicial

opcao :: Opcoes

nivel :: Int

data Imagem

Constructors

BlocoPlataforma

BlocoAlcapao

PersJogador

PersFantasma

PersMacacoMalvado

ColMoeda

ColMartelo

Estrela

BlocoEscada

MJ

Escolher

AutoriaM

MenuBG

DesertoM

LagoaM

IlhasM

PortalM

Autoras

Voltar

VidaF

DesertoBG

LagoaBG

IlhasBG

PortalBG

JogadorArmado

GameOver

Victory

Nome

type Imagens = [(Imagem, Picture)]

data MenuInicial

Constructors

MenuPrincipal

EscolherNivelMenu

PlayerMenu

CreditosMenu

data Opcoes

Constructors

Jogar

EscolherNivel

HistoricoPlayers

Creditos

Mapa1

Mapa2

Mapa3

Mapa4

VoltarMenuIncial

data Modo

Constructors

MenuInicial

EmJogo

Vitoria

Derrota

type PontosJogador = [(String, Int)]

altura :: Float

Retorna a altura da tela do jogo.

comprimento :: Float

Retorna o comprimento da tela do jogo.

tamanhoMatriz :: Mapa -> (Int, Int)

Recebe um mapa e retorna um par representando as dimensões da matriz do mapa.

larguraBloco :: Mapa -> Double

Recebe um mapa e retorna a largura de um bloco no jogo.

alturaBloco :: Mapa -> Double

Recebe um mapa e retorna a altura de um bloco no jogo

loadImagens :: IO Imagens

Carrega imagens em formato BMP, utilizando a função loadBMP para carregar cada imagem associada a uma determinada chave (identificada pelos construtores de dados)

vazio :: Picture

getImagem :: Imagem -> Imagens -> Picture

Recebe uma Imagem e um dicionário de Imagens, onde uma Imagem é mapeada para uma Picture

desenhaBloco :: Bloco -> Imagens -> Picture

Recebe um Bloco e um dicionário de Imagens. Com base no tipo de Bloco, ela utiliza a função getImagem para obter a Picture correspondente a esse tipo no dicionário de imagens.

desenhaMapa :: Mapa -> Imagens -> Picture

Utiliza a função blocosPosicaoMatriz para obter uma lista de tuplas contendo o tipo de bloco e sua posição na matriz do mapa. Para cada tupla, a função cria uma Picture correspondente ao bloco, utilizando a função desenhaBloco e usa a função pictures para compor todas essas imagens juntas. Cada Picture é criada com base nas coordenadas do bloco na tela (translate) e na escala definida para a largura e altura do bloco (scale). A escala é ajustada para garantir que as imagens tenham um tamanho proporcional ao tamanho padrão (50x50) esperado para os blocos.

desenhaJogador :: Estado -> Imagens -> Picture

Desenha o jogador na tela. Se o jogador está morto, retorna uma imagem em branco (Blank).

desenhaColecionaveis :: Jogo -> Imagens -> Picture

Desenha os colecionáveis na tela. Se a lista de colecionáveis no jogo for vazia, ela retorna uma imagem em branco (Blank). Senão, cria uma lista de imagens, cada uma representando um colecionável, e posiciona-as de acordo com suas coordenadas na matriz do jogo.

desenhaColecionavel :: (Colecionavel, Posicao) -> Imagens -> Picture

Seleciona a imagem correspondente ao tipo de colecionável (Martelo ou Moeda) e dimensiona-na.

desenhaInimigos :: Jogo -> Imagens -> Picture

Desenha todos os inimigos presentes no jogo. Se não houver inimigos, retorna uma imagem em branco (Blank). Caso contrário,utiliza a função desenhaInimigo para desenhar cada inimigo individualmente.

desenhaInimigo :: Personagem -> Jogo -> Imagens -> Picture

Devolver imagem que representa o inimigo na posição correta do jogo, com escala e direção ajustadas. Se a vida do inimigo (vida i) for maior que 0, decide qual tipo de inimigo é (Fantasma ou Macaco Malvado) e ajusta a escala e a direção. Caso contrário, retorna uma imagem em branco (BlanvalidaEscadas (Mapa ((0.5, 5.5), Oeste) (0.5, 2.5) [[Escada, Escada, Escada], [Vazio, Vazio, Vazio], [Plataforma, Plataforma, Plataforma]]) k), indicando que o inimigo está morto.

desenhaJogo :: Estado -> Imagens -> Picture

Dependendo do modo de jogo no estado, decide o que desenhar na tela. Se o modo for EmJogo, desenha vários elementos do jogo. Caso contrário, desenha o menu principal.

desenhaInfo :: Personagem -> Imagens -> Picture

Cria uma imagem composta por uma imagem da barra de vida e por um texto que exibe informações sobre o jogador, como pontuação e número de vidas.

desenhaEstrela :: Mapa -> Imagens -> Picture

Desenha a estrela (chave) na tela

desenharVitoria :: Imagens -> Picture

Desenhar a caixa de vitória na tela.

desenharDerrota :: Imagens -> Picture

Desenhar a caixa de derrota na tela.

posicaoEcra :: Mapa -> Posicao -> (Float, Float)

Converte uma posição no mapa para as coordenadas correspondentes na tela.

desenhaMenuPrincipal :: Estado -> Picture

Desenha os elementos do menu principal com base no estado do jogo. A opção selecionada é destacada com um tamanho maior.

desenhaFundo :: Estado -> Imagens -> Picture

Desenha o fundo do jogo com base no nível atual.

mapear :: Event -> Estado -> Estado

Mapeamento de eventos para ações no jogo.

reageTempo :: Float -> Estado -> Estado

Reage ao tempo no estado do jogo durante o modo EmJogo.

chegadaEstrela :: Personagem -> Mapa -> Bool

Verifica se o jogador atingiu a estrela num determinado mapa.

iniciarJogoNoNivel :: Estado -> Int -> Estado

Recebe um Estado e um número de nível escolhido. Retorna um novo Estado inicializado para o nível especificado.

estadoInicial :: Imagens -> Estado -> Estado

Retorna um novo estado inicializado de acordo com o nível especificado no estado original

estadoInicialMenu :: Imagens -> Estado

Retorna um novo estado inicializado para o menu principal. Não recebe um estado existente como argumento, apenas um conjunto de imagens (Imagens).

mapa01 :: Mapa

mapa02 :: Mapa

mapa03 :: Mapa

mapa04 :: Mapa

jogoInicial :: Jogo

jogo2 :: Jogo

jogo3 :: Jogo

jogo4 :: Jogo

acoes :: [Personagem] -> [Maybe Acao]

jog :: Personagem

c1 :: (Colecionavel, (Double, Double))

c2 :: (Colecionavel, (Double, Double))

c3 :: (Colecionavel, (Double, Double))

c4 :: (Colecionavel, (Double, Double))

c5 :: (Colecionavel, (Double, Double))

c6 :: (Colecionavel, (Double, Double))

c7 :: (Colecionavel, (Double, Double))

c8 :: (Colecionavel, (Double, Double))

en1 :: Personagem

en2 :: Personagem

en3 :: Personagem

en4 :: Personagem

en5 :: Personagem

en6 :: Personagem

en7 :: Personagem

en8 :: Personagem